refactor: 拆分 config.mjs 的 sidebar 数据与构建钩子到独立模块
docs/.vitepress/config.mjs 从 4698 行降至 987 行（-79%）：

- sidebars/data.mjs: 静态 sidebar 与多语言 labels 数据
- sidebars/index.mjs: sidebar 生成/本地化函数
- seo.mjs: getSeoHead 与 locale 链接工具（工厂模式注入共享配置）
- build-hooks.mjs: transformHtml/transformHead/buildEnd 构建钩子

行为保持不变：10 种语言（含 root）的 head/nav/sidebar/docFooter
等 130 项配置输出与拆分前逐项对比一致。
