+++ docs/.vitepress/sidebars/data.mjs
@@ -1,0 +1,3132 @@
+// Static sidebar & label data for the VitePress site.
+// Extracted from docs/.vitepress/config.mjs to keep the main config lean.
+
+export const productManagerSidebarEn = [
+  {
+    text: 'Chapter 1: Introduction to Programming',
+    collapsed: false,
+    items: [
+      { text: 'How to Learn This Course', link: '/en/stage-1/learning-map/' },
+      {
+        text: 'First Programming Experience in the AI Era',
+        link: '/en/stage-1/ai-capabilities-through-games/'
+      }
+    ]
+  },
+  {
+    text: 'Chapter 2: Development Environment and Tools',
+    collapsed: false,
+    items: [
+      {
+        text: 'Introduction to AI Programming Tools',
+        link: '/en/stage-1/introduction-to-ai-ide/'
+      },
+      {
+        text: 'Clone from Screenshots: First Imitation Practice',
+        link: '/en/stage-1/clone-your-favorite-app/'
+      }
+    ]
+  },
+  {
+    text: 'Chapter 3: From Problems to Solutions',
+    collapsed: false,
+    items: [
+      {
+        text: '1. Find Real Problems',
+        link: '/en/stage-1/appendix-idea-sources/'
+      },
+      {
+        text: '2. Select a Direction Worth Building',
+        link: '/en/stage-1/finding-great-idea/'
+      },
+      {
+        text: '3. Understand Real Needs (JTBD)',
+        link: '/en/stage-1/appendix-jobs-to-be-done/'
+      },
+      {
+        text: '4. Validate Needs Through Interviews',
+        link: '/en/stage-1/appendix-mom-test/'
+      },
+      {
+        text: '5. Narrow the Solution (Double Diamond)',
+        link: '/en/stage-1/appendix-double-diamond/'
+      }
+    ]
+  },
+  {
+    text: 'Chapter 4: From Solution to Product',
+    collapsed: false,
+    items: [
+      {
+        text: '6. Build an Interactive Prototype',
+        link: '/en/stage-1/building-prototype/'
+      },
+      {
+        text: '7. Connect Real AI Capabilities',
+        link: '/en/stage-1/integrating-ai-capabilities/'
+      },
+      {
+        text: '8. Polish and Deliver a Complete Product',
+        link: '/en/stage-1/complete-project-practice/'
+      }
+    ]
+  },
+  {
+    text: 'Appendix',
+    collapsed: false,
+    items: [
+      {
+        text: 'Publish Your Vibe Coding Project on ModelScope',
+        link: '/en/stage-1/appendix-modelscope-static-site/'
+      },
+      {
+        text: 'Product Thinking Fundamentals',
+        link: '/en/stage-1/appendix-a-product-thinking/'
+      },
+      {
+        text: 'Find AI Scenarios in Real Workflows',
+        link: '/en/stage-1/appendix-industry-scenarios/'
+      },
+      {
+        text: 'Common Issues and Troubleshooting',
+        link: '/en/stage-1/appendix-b-common-errors/'
+      }
+    ]
+  }
+]
+
+export const productManagerSidebarKo = [
+  {
+    text: '1장 프로그래밍 입문',
+    collapsed: false,
+    items: [
+      { text: '이 과목 학습 방법', link: '/ko-kr/stage-1/learning-map/' },
+      {
+        text: 'AI 시대의 프로그래밍 첫 경험',
+        link: '/ko-kr/stage-1/ai-capabilities-through-games/'
+      }
+    ]
+  },
+  {
+    text: '2장 개발 환경과 도구',
+    collapsed: false,
+    items: [
+      {
+        text: 'AI 프로그래밍 도구 소개와 사용',
+        link: '/ko-kr/stage-1/introduction-to-ai-ide/'
+      },
+      {
+        text: '스크린샷으로 복제하기: 첫 모방 연습',
+        link: '/ko-kr/stage-1/clone-your-favorite-app/'
+      }
+    ]
+  },
+  {
+    text: '3장 문제에서 해결책까지',
+    collapsed: false,
+    items: [
+      {
+        text: '1. 실제 문제 찾기',
+        link: '/ko-kr/stage-1/appendix-idea-sources/'
+      },
+      {
+        text: '2. 만들 가치가 있는 방향 고르기',
+        link: '/ko-kr/stage-1/finding-great-idea/'
+      },
+      {
+        text: '3. 실제 요구 이해하기(JTBD)',
+        link: '/ko-kr/stage-1/appendix-jobs-to-be-done/'
+      },
+      {
+        text: '4. 인터뷰로 요구 검증하기',
+        link: '/ko-kr/stage-1/appendix-mom-test/'
+      },
+      {
+        text: '5. 해결책 좁히기(더블 다이아몬드)',
+        link: '/ko-kr/stage-1/appendix-double-diamond/'
+      }
+    ]
+  },
+  {
+    text: '4장 해결책에서 작품까지',
+    collapsed: false,
+    items: [
+      {
+        text: '6. 인터랙티브 프로토타입 만들기',
+        link: '/ko-kr/stage-1/building-prototype/'
+      },
+      {
+        text: '7. 실제 AI 기능 연결하기',
+        link: '/ko-kr/stage-1/integrating-ai-capabilities/'
+      },
+      {
+        text: '8. 다듬고 완성된 작품으로 제출하기',
+        link: '/ko-kr/stage-1/complete-project-practice/'
+      }
+    ]
+  },
+  {
+    text: '부록',
+    collapsed: false,
+    items: [
+      {
+        text: 'Vibe Coding 결과물을 ModelScope에 게시하기',
+        link: '/ko-kr/stage-1/appendix-modelscope-static-site/'
+      },
+      {
+        text: '제품 사고 기초',
+        link: '/ko-kr/stage-1/appendix-a-product-thinking/'
+      },
+      {
+        text: '실제 업무 흐름에서 AI 장면 찾기',
+        link: '/ko-kr/stage-1/appendix-industry-scenarios/'
+      },
+      {
+        text: '자주 묻는 질문과 문제 해결',
+        link: '/ko-kr/stage-1/appendix-b-common-errors/'
+      }
+    ]
+  }
+]
+
+export const stage2SidebarEn = [
+  {
+    text: 'Frontend Development',
+    collapsed: false,
+    items: [
+      {
+        text: 'Asset Production & Agent Building',
+        link: '/en/stage-2/frontend/lovart-assets/'
+      },
+      {
+        text: 'AI Design Basics',
+        link: '/en/stage-2/frontend/ui-design/'
+      },
+      {
+        text: 'Figma & MasterGo Basics',
+        link: '/en/stage-2/frontend/figma-mastergo/'
+      },
+      {
+        text: 'UI Design Specs & Multi-Product Interface',
+        link: '/en/stage-2/frontend/multi-product-ui/'
+      },
+      {
+        text: 'LLM & Skills Interface Beautification',
+        link: '/en/stage-2/frontend/llm-skills-beautiful/'
+      },
+      {
+        text: 'Design Prototype to Project Code',
+        link: '/en/stage-2/frontend/design-to-code/'
+      },
+      {
+        text: 'Modern Component Libraries',
+        link: '/en/stage-2/frontend/modern-component-library/'
+      }
+    ]
+  },
+  {
+    text: 'Backend & Full-Stack',
+    collapsed: false,
+    items: [
+      {
+        text: 'Database & Supabase',
+        link: '/en/stage-2/backend/database-supabase/'
+      },
+      {
+        text: 'Backend API Design & Development',
+        link: '/en/stage-2/backend/ai-interface-code/'
+      },
+      {
+        text: 'Git & GitHub Workflow',
+        link: '/en/stage-2/backend/git-workflow/'
+      },
+      {
+        text: 'Put Your Website Online (Easy): One-Click PaaS (Vercel/Zeabur)',
+        link: '/en/stage-2/backend/zeabur-deployment/'
+      },
+      {
+        text: 'Put Your Website Online (Advanced): Set Up Your Own VPS',
+        link: '/en/stage-2/backend/cloud-server-deployment/'
+      },
+      {
+        text: 'Modern CLI Dev Tools',
+        link: '/en/stage-2/backend/modern-cli/'
+      },
+      {
+        text: 'Stripe Payment Integration',
+        link: '/en/stage-2/backend/stripe-payment/'
+      }
+    ]
+  },
+  {
+    text: 'AI Capabilities Appendix',
+    collapsed: false,
+    items: [
+      {
+        text: 'AI 1: Dify & Knowledge Base',
+        link: '/en/stage-2/ai-capabilities/dify-knowledge-base/'
+      }
+    ]
+  },
+  {
+    text: 'Comprehensive Projects',
+    collapsed: false,
+    items: [
+      {
+        text: 'Hogwarts Portraits Project',
+        link: '/en/stage-2/frontend/hogwarts-portraits/'
+      },
+      {
+        text: 'AI Marketing Copywriting SaaS',
+        link: '/en/stage-2/assignments/copywriting-platform-supabase/'
+      },
+      {
+        text: 'Online Exam & Management System',
+        link: '/en/stage-2/assignments/exam-management-express/'
+      },
+      {
+        text: 'Modern AI Image SaaS',
+        link: '/en/stage-2/assignments/modern-landing-page/'
+      },
+      {
+        text: 'Custom Dify Agent Platform',
+        link: '/en/stage-2/assignments/custom-dify-agent-platform/'
+      },
+      {
+        text: 'Travel Planning Agent Platform',
+        link: '/en/stage-2/assignments/travel-planning-agent-platform/'
+      },
+      {
+        text: 'Spring Boot Movie Recommender',
+        link: '/en/stage-2/assignments/movie-recommendation-springboot/'
+      },
+      {
+        text: 'Grocery Microservices System',
+        link: '/en/stage-2/assignments/simple-grocery-microservices/'
+      },
+      {
+        text: 'Go Traffic Data Platform',
+        link: '/en/stage-2/assignments/traffic-data-visualization-go/'
+      }
+    ]
+  }
+]
+
+export const zhCnStage2Sidebar = [
+  {
+    text: '前端开发',
+    collapsed: false,
+    items: [
+      {
+        text: 'NanoBanana 素材生产',
+        link: '/zh-cn/stage-2/frontend/lovart-assets/'
+      },
+      {
+        text: 'AI 设计入门',
+        link: '/zh-cn/stage-2/frontend/ui-design/'
+      },
+      {
+        text: 'Figma 与 MasterGo 入门',
+        link: '/zh-cn/stage-2/frontend/figma-mastergo/'
+      },
+      {
+        text: 'UI 设计规范与多产品界面',
+        link: '/zh-cn/stage-2/frontend/multi-product-ui/'
+      },
+      {
+        text: '结合 Agent Skills 美化界面',
+        link: '/zh-cn/stage-2/frontend/llm-skills-beautiful/'
+      },
+      {
+        text: '设计原型到项目代码',
+        link: '/zh-cn/stage-2/frontend/design-to-code/'
+      },
+      {
+        text: '现代组件库与界面升级',
+        link: '/zh-cn/stage-2/frontend/modern-component-library/'
+      }
+    ]
+  },
+  {
+    text: '后端开发',
+    collapsed: false,
+    items: [
+      {
+        text: '数据库与 Supabase 入门',
+        link: '/zh-cn/stage-2/backend/database-supabase/'
+      },
+      {
+        text: '大模型辅助接口开发',
+        link: '/zh-cn/stage-2/backend/ai-interface-code/'
+      },
+      {
+        text: 'Git 与 GitHub 入门指南',
+        link: '/zh-cn/stage-2/backend/git-workflow/'
+      },
+      {
+        text: '把网站发到网上（简单方式）：Vercel/Zeabur/CloudBase 一键发布',
+        link: '/zh-cn/stage-2/backend/zeabur-deployment/'
+      },
+      {
+        text: '把网站发到网上（进阶方式）：自己买台 VPS 服务器搭建发布',
+        link: '/zh-cn/stage-2/backend/cloud-server-deployment/'
+      },
+      {
+        text: 'CLI Coding Agent 编程助手',
+        link: '/zh-cn/stage-2/backend/modern-cli/'
+      },
+      {
+        text: 'Stripe 支付集成',
+        link: '/zh-cn/stage-2/backend/stripe-payment/'
+      }
+    ]
+  },
+  {
+    text: 'AI 能力附录',
+    collapsed: false,
+    items: [
+      {
+        text: 'Dify 入门与知识库集成',
+        link: '/zh-cn/stage-2/ai-capabilities/dify-knowledge-base/'
+      }
+    ]
+  },
+  {
+    text: '综合项目',
+    collapsed: false,
+    items: [
+      {
+        text: '一起做霍格沃茨画像',
+        link: '/zh-cn/stage-2/frontend/hogwarts-portraits/'
+      },
+      {
+        text: 'AI 营销文案 SaaS',
+        link: '/zh-cn/stage-2/assignments/copywriting-platform-supabase/'
+      },
+      {
+        text: '在线考试与管理系统',
+        link: '/zh-cn/stage-2/assignments/exam-management-express/'
+      },
+      {
+        text: '现代 AI 生图 SaaS',
+        link: '/zh-cn/stage-2/assignments/modern-landing-page/'
+      },
+      {
+        text: '类 Dify 智能体平台',
+        link: '/zh-cn/stage-2/assignments/custom-dify-agent-platform/'
+      },
+      {
+        text: '智能旅游规划 Agent 平台',
+        link: '/zh-cn/stage-2/assignments/travel-planning-agent-platform/'
+      },
+      {
+        text: 'Spring Boot 电影推荐系统',
+        link: '/zh-cn/stage-2/assignments/movie-recommendation-springboot/'
+      },
+      {
+        text: '生鲜电商微服务系统',
+        link: '/zh-cn/stage-2/assignments/simple-grocery-microservices/'
+      },
+      {
+        text: 'Go 交通数据分析平台',
+        link: '/zh-cn/stage-2/assignments/traffic-data-visualization-go/'
+      }
+    ]
+  }
+]
+
+export const stage3SidebarEn = [
+  {
+    text: 'Core Skills',
+    collapsed: false,
+    items: [
+      {
+        text: 'Claude Code Quickstart Core Guide',
+        link: '/en/stage-3/core-skills/basics/'
+      },
+      {
+        text: 'MCP and Claude Code Complete Guide',
+        link: '/en/stage-3/core-skills/mcp/'
+      },
+      {
+        text: 'Claude Code Skills Complete Guide',
+        link: '/en/stage-3/core-skills/skills/'
+      },
+      {
+        text: 'Making Claude Code Work on Long-Running Tasks',
+        link: '/en/stage-3/core-skills/long-running-tasks/'
+      },
+      {
+        text: 'Claude Agent Teams Complete Guide',
+        link: '/en/stage-3/core-skills/agent-teams/'
+      },
+      {
+        text: 'AI-Assisted Development Workflow',
+        link: '/en/stage-3/core-skills/workflow/'
+      },
+      {
+        text: 'Claude Code Remote Development on Mobile',
+        link: '/en/stage-3/core-skills/mobile-development/'
+      },
+      {
+        text: 'Claude Agent SDK Complete Guide',
+        link: '/en/stage-3/core-skills/claude-agent-sdk/'
+      }
+    ]
+  },
+  {
+    text: 'Specification-Driven Development',
+    collapsed: false,
+    items: [
+      {
+        text: 'Claude Code Superpowers for Production-Grade Development',
+        link: '/en/stage-3/core-skills/superpowers/'
+      },
+      {
+        text: 'From Vibe Coding to Spec Coding',
+        link: '/en/stage-3/core-skills/spec-coding/'
+      },
+      {
+        text: 'Building Software End to End with GitHub Issues',
+        link: '/en/stage-3/core-skills/github-iterative-development/'
+      }
+    ]
+  },
+  {
+    text: 'Cross-Platform Development',
+    collapsed: false,
+    items: [
+      {
+        text: 'From a Web Page to Real Software and an AI-Native Product',
+        link: '/en/stage-3/cross-platform/ai-native-creator/'
+      },
+      {
+        text: 'How to Choose the Right Platform for Your App',
+        link: '/en/stage-3/cross-platform/choose-platform/'
+      },
+      {
+        text: 'How to Build a Simple WeChat Mini Program',
+        link: '/en/stage-3/cross-platform/wechat-miniprogram/'
+      },
+      {
+        text: 'How to Build a WeChat Mini Program with a Backend',
+        link: '/en/stage-3/cross-platform/wechat-miniprogram-backend/'
+      },
+      {
+        text: 'How to Build an Android App with Jetpack Compose',
+        link: '/en/stage-3/cross-platform/android-app/'
+      },
+      {
+        text: 'How to Build an iOS App with SwiftUI',
+        link: '/en/stage-3/cross-platform/ios-app/'
+      },
+      {
+        text: 'Build a Store Inspection App with React Native and Expo',
+        link: '/en/stage-3/cross-platform/react-native-expo/'
+      },
+      {
+        text: 'How to Build a Cross-Platform App with Flutter',
+        link: '/en/stage-3/cross-platform/flutter-app/'
+      },
+      {
+        text: 'How to Build a PWA Local App',
+        link: '/en/stage-3/cross-platform/pwa-local-app/'
+      },
+      {
+        text: 'How to Build a Browser AI Assistant Extension',
+        link: '/en/stage-3/cross-platform/browser-ai-extension/'
+      },
+      {
+        text: 'How to Build a Cross-Platform Electron Desktop App',
+        link: '/en/stage-3/cross-platform/electron-voice-to-text/'
+      },
+      {
+        text: 'How to Build Platform, Pixel-Art, and 3D Games with Godot',
+        link: '/en/stage-3/cross-platform/godot-game-development/'
+      },
+      {
+        text: 'How to Quickly Build and Mint an NFT',
+        link: '/en/stage-3/cross-platform/nft-minting/'
+      },
+      {
+        text: 'How to Build a VS Code Extension',
+        link: '/en/stage-3/cross-platform/vscode-extension/'
+      },
+      {
+        text: 'How to Build an Industrial Qt Desktop App',
+        link: '/en/stage-3/cross-platform/qt-industrial-hmi/'
+      },
+      {
+        text: 'How to Build Your Personal Website and Academic Blog',
+        link: '/en/stage-3/personal-brand/personal-website-blog/'
+      },
+      {
+        text: 'How to Publish the Application You Built',
+        link: '/en/stage-3/cross-platform/app-publishing/'
+      }
+    ]
+  },
+  {
+    text: 'AI Advanced',
+    collapsed: false,
+    items: [
+      {
+        text: 'What Is RAG and How It Works',
+        link: '/en/stage-3/ai-advanced/rag-introduction/'
+      },
+      {
+        text: 'Advanced RAG and Workflow Orchestration with LangGraph',
+        link: '/en/stage-3/ai-advanced/langgraph-advanced-rag/'
+      }
+    ]
+  }
+]
+
+export const appendixSidebarEn = [
+  {
+    text: 'I. Computer Fundamentals',
+    collapsed: false,
+    items: [
+      {
+        text: 'Full-Stack Development: An Introduction',
+        link: '/zh-cn/appendix/1-computer-fundamentals/vibe-coding-fullstack'
+      },
+      {
+        text: 'Computer System Overview: From Boot to Web',
+        link: '/zh-cn/appendix/1-computer-fundamentals/power-on-to-web'
+      },
+      {
+        text: 'Digital Circuits Basics: From Transistors to CPU',
+        link: '/zh-cn/appendix/1-computer-fundamentals/transistor-to-cpu'
+      },
+      {
+        text: 'Computer Organization Principles',
+        link: '/zh-cn/appendix/1-computer-fundamentals/computer-organization'
+      },
+      {
+        text: 'Operating Systems Principles',
+        link: '/zh-cn/appendix/1-computer-fundamentals/operating-systems'
+      },
+      {
+        text: 'Data Representation: Encoding, Storage, Transmission',
+        link: '/zh-cn/appendix/1-computer-fundamentals/data-encoding-storage'
+      },
+      {
+        text: 'Computer Networks: From URL to Rendering',
+        link: '/zh-cn/appendix/1-computer-fundamentals/computer-networks'
+      },
+      {
+        text: 'Data Structures: An Introduction',
+        link: '/zh-cn/appendix/1-computer-fundamentals/data-structures'
+      },
+      {
+        text: 'Introduction to Algorithms',
+        link: '/zh-cn/appendix/1-computer-fundamentals/algorithm-thinking'
+      },
+      {
+        text: 'Programming Language Concepts',
+        link: '/zh-cn/appendix/1-computer-fundamentals/programming-languages'
+      },
+      {
+        text: 'Compiler Principles',
+        link: '/zh-cn/appendix/1-computer-fundamentals/compilers'
+      },
+      {
+        text: 'Type Systems: An Introduction',
+        link: '/zh-cn/appendix/1-computer-fundamentals/type-systems'
+      }
+    ]
+  },
+  {
+    text: 'II. Tools & Environment',
+    collapsed: false,
+    items: [
+      {
+        text: 'IDE Basics',
+        link: '/en/appendix/2-development-tools/ide-basics'
+      },
+      {
+        text: 'Command Line & Shell',
+        link: '/zh-cn/appendix/2-development-tools/command-line-shell'
+      },
+      {
+        text: 'Git: Code Time Machine',
+        link: '/zh-cn/appendix/2-development-tools/git-version-control'
+      },
+      {
+        text: 'Env Vars & PATH',
+        link: '/zh-cn/appendix/2-development-tools/environment-path'
+      },
+      {
+        text: 'Ports & Localhost',
+        link: '/zh-cn/appendix/2-development-tools/ports-localhost'
+      },
+      {
+        text: 'SSH & Key Auth',
+        link: '/zh-cn/appendix/2-development-tools/ssh-authentication'
+      },
+      {
+        text: 'Package Managers',
+        link: '/zh-cn/appendix/2-development-tools/package-managers'
+      },
+      {
+        text: 'Art of Debugging',
+        link: '/zh-cn/appendix/2-development-tools/debugging-art'
+      },
+      {
+        text: 'Regex',
+        link: '/zh-cn/appendix/2-development-tools/regex'
+      }
+    ]
+  },
+  {
+    text: 'III. Browser & Frontend',
+    collapsed: false,
+    items: [
+      {
+        text: 'JavaScript Deep Dive',
+        link: '/zh-cn/appendix/3-browser-and-frontend/javascript-deep-dive'
+      },
+      {
+        text: 'TypeScript Intro',
+        link: '/zh-cn/appendix/3-browser-and-frontend/typescript'
+      },
+      {
+        text: 'Frontend Frameworks',
+        link: '/zh-cn/appendix/3-browser-and-frontend/frontend-frameworks'
+      },
+      {
+        text: 'Rendering Pipeline',
+        link: '/zh-cn/appendix/3-browser-and-frontend/browser-as-os-rendering'
+      },
+      {
+        text: 'HTML / CSS Layout',
+        link: '/zh-cn/appendix/3-browser-and-frontend/html-css-layout'
+      },
+      {
+        text: 'JS Runtime',
+        link: '/zh-cn/appendix/3-browser-and-frontend/javascript-runtime'
+      },
+      {
+        text: 'Nature of Frameworks',
+        link: '/zh-cn/appendix/3-browser-and-frontend/frontend-framework-nature'
+      },
+      {
+        text: 'State Management',
+        link: '/zh-cn/appendix/3-browser-and-frontend/state-management'
+      },
+      {
+        text: 'Routing & Navigation',
+        link: '/zh-cn/appendix/3-browser-and-frontend/routing-navigation'
+      },
+      {
+        text: 'Graphics & Animation',
+        link: '/zh-cn/appendix/3-browser-and-frontend/graphics-animation'
+      },
+      {
+        text: 'Real-time Comm',
+        link: '/zh-cn/appendix/3-browser-and-frontend/realtime-communication'
+      },
+      {
+        text: 'Web Performance',
+        link: '/zh-cn/appendix/3-browser-and-frontend/web-performance'
+      },
+      {
+        text: 'Frontend Engineering',
+        link: '/zh-cn/appendix/3-browser-and-frontend/frontend-engineering'
+      },
+      {
+        text: 'Project Architecture',
+        link: '/zh-cn/appendix/3-browser-and-frontend/frontend-project-architecture'
+      },
+      {
+        text: 'A11y & i18n',
+        link: '/zh-cn/appendix/3-browser-and-frontend/a11n-i18n'
+      }
+    ]
+  },
+  {
+    text: 'IV. Server & Backend',
+    collapsed: false,
+    items: [
+      {
+        text: 'Backend Languages',
+        link: '/zh-cn/appendix/4-server-and-backend/backend-languages'
+      },
+      {
+        text: 'Client Languages',
+        link: '/zh-cn/appendix/4-server-and-backend/client-languages'
+      },
+      {
+        text: 'Cross-platform Solutions',
+        link: '/zh-cn/appendix/4-server-and-backend/cross-platform'
+      },
+      {
+        text: 'HTTP Protocol',
+        link: '/zh-cn/appendix/4-server-and-backend/http-protocol'
+      },
+      {
+        text: 'Request Journey',
+        link: '/zh-cn/appendix/4-server-and-backend/request-journey'
+      },
+      {
+        text: 'Web Frameworks',
+        link: '/zh-cn/appendix/4-server-and-backend/web-frameworks'
+      },
+      {
+        text: 'API Intro',
+        link: '/zh-cn/appendix/4-server-and-backend/api-intro'
+      },
+      {
+        text: 'API Design Philosophy',
+        link: '/zh-cn/appendix/4-server-and-backend/api-design'
+      },
+      {
+        text: 'Serialization',
+        link: '/zh-cn/appendix/4-server-and-backend/serialization'
+      },
+      {
+        text: 'Auth & Authorization',
+        link: '/zh-cn/appendix/4-server-and-backend/auth-authorization'
+      },
+      {
+        text: 'Concurrency & Async',
+        link: '/zh-cn/appendix/4-server-and-backend/concurrency-async'
+      },
+      {
+        text: 'Caching Strategies',
+        link: '/zh-cn/appendix/4-server-and-backend/caching'
+      },
+      {
+        text: 'Message Queues',
+        link: '/zh-cn/appendix/4-server-and-backend/message-queues'
+      },
+      {
+        text: 'Async Task Queues',
+        link: '/zh-cn/appendix/4-server-and-backend/async-task-queues'
+      },
+      {
+        text: 'Rate Limiting',
+        link: '/zh-cn/appendix/4-server-and-backend/rate-limiting-backpressure'
+      },
+      {
+        text: 'Search Engine Principles',
+        link: '/zh-cn/appendix/4-server-and-backend/search-engines'
+      },
+      {
+        text: 'File Storage',
+        link: '/zh-cn/appendix/4-server-and-backend/file-storage'
+      },
+      {
+        text: 'Backend Architecture',
+        link: '/zh-cn/appendix/4-server-and-backend/backend-layered-architecture'
+      },
+      {
+        text: 'Project Architecture',
+        link: '/zh-cn/appendix/4-server-and-backend/backend-project-architecture'
+      },
+      {
+        text: 'DSL Intro',
+        link: '/zh-cn/appendix/4-server-and-backend/domain-specific-languages'
+      }
+    ]
+  },
+  {
+    text: 'V. Data',
+    collapsed: false,
+    items: [
+      {
+        text: 'Database Fundamentals',
+        link: '/zh-cn/appendix/5-data/database-fundamentals'
+      },
+      {
+        text: 'Data Models Panorama',
+        link: '/zh-cn/appendix/5-data/data-models'
+      },
+      {
+        text: 'Data Tracking',
+        link: '/zh-cn/appendix/5-data/data-tracking'
+      },
+      {
+        text: 'Data Analysis',
+        link: '/zh-cn/appendix/5-data/data-analysis'
+      },
+      {
+        text: 'A/B Testing',
+        link: '/zh-cn/appendix/5-data/ab-testing'
+      },
+      {
+        text: 'Data Visualization',
+        link: '/zh-cn/appendix/5-data/data-visualization'
+      },
+      {
+        text: 'Data Governance',
+        link: '/zh-cn/appendix/5-data/data-governance'
+      }
+    ]
+  },
+  {
+    text: 'VI. Architecture',
+    collapsed: false,
+    items: [
+      {
+        text: 'Monolith to Microservices',
+        link: '/zh-cn/appendix/6-architecture-and-system-design/monolith-to-microservices'
+      },
+      {
+        text: 'Distributed Systems',
+        link: '/zh-cn/appendix/6-architecture-and-system-design/distributed-systems'
+      },
+      {
+        text: 'HA & Disaster Recovery',
+        link: '/zh-cn/appendix/6-architecture-and-system-design/high-availability'
+      },
+      {
+        text: 'System Design',
+        link: '/zh-cn/appendix/6-architecture-and-system-design/system-design-methodology'
+      }
+    ]
+  },
+  {
+    text: 'VII. Infrastructure',
+    collapsed: false,
+    items: [
+      {
+        text: 'Linux Basics',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/linux-basics'
+      },
+      {
+        text: 'Docker Containers',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/docker-containers'
+      },
+      {
+        text: 'Kubernetes',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/kubernetes'
+      },
+      {
+        text: 'CI / CD',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/ci-cd'
+      },
+      {
+        text: 'Domain, DNS & HTTPS',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/dns-https'
+      },
+      {
+        text: 'Load Balancing',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/load-balancing-gateway'
+      },
+      {
+        text: 'Gateway & Reverse Proxy',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/gateway-proxy'
+      },
+      {
+        text: 'Cloud Platforms',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-platforms'
+      },
+      {
+        text: 'IAM',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-iam'
+      },
+      {
+        text: 'Storage & CDN',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-storage-cdn'
+      },
+      {
+        text: 'IaC',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/infrastructure-as-code'
+      },
+      {
+        text: 'Monitoring & Logging',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/monitoring-logging'
+      },
+      {
+        text: 'Incident Response',
+        link: '/zh-cn/appendix/7-infrastructure-and-operations/incident-response'
+      }
+    ]
+  },
+  {
+    text: 'VIII. Artificial Intelligence',
+    collapsed: false,
+    items: [
+      {
+        text: 'AI History & Concepts',
+        link: '/en/appendix/8-artificial-intelligence/ai-history'
+      },
+      {
+        text: 'Neural Networks',
+        link: '/zh-cn/appendix/8-artificial-intelligence/neural-networks'
+      },
+      {
+        text: 'Transformer & Attention',
+        link: '/zh-cn/appendix/8-artificial-intelligence/transformer-attention'
+      },
+      {
+        text: 'LLM Principles',
+        link: '/zh-cn/appendix/8-artificial-intelligence/llm-principles'
+      },
+      {
+        text: 'Prompt Engineering',
+        link: '/zh-cn/appendix/8-artificial-intelligence/prompt-engineering'
+      },
+      {
+        text: 'Context Engineering',
+        link: '/zh-cn/appendix/8-artificial-intelligence/context-engineering'
+      },
+      {
+        text: 'Multimodal Models',
+        link: '/zh-cn/appendix/8-artificial-intelligence/multimodal-models'
+      },
+      {
+        text: 'Image Generation',
+        link: '/zh-cn/appendix/8-artificial-intelligence/image-generation'
+      },
+      {
+        text: 'Speech Synthesis',
+        link: '/zh-cn/appendix/8-artificial-intelligence/speech-synthesis-recognition'
+      },
+      {
+        text: 'Embedding & Vector Search',
+        link: '/zh-cn/appendix/8-artificial-intelligence/embedding-vector-retrieval'
+      },
+      {
+        text: 'RAG Architecture',
+        link: '/zh-cn/appendix/8-artificial-intelligence/rag'
+      },
+      {
+        text: 'AI Agent & Tools',
+        link: '/zh-cn/appendix/8-artificial-intelligence/ai-agents'
+      },
+      {
+        text: 'AI Protocols (MCP)',
+        link: '/zh-cn/appendix/8-artificial-intelligence/ai-protocols'
+      },
+      {
+        text: 'Fine-tuning & Deployment',
+        link: '/zh-cn/appendix/8-artificial-intelligence/model-finetuning-deployment'
+      },
+      {
+        text: 'AI Native Design',
+        link: '/zh-cn/appendix/8-artificial-intelligence/ai-native-app-design'
+      },
+      {
+        text: 'AI Dictionary',
+        link: '/zh-cn/appendix/8-artificial-intelligence/ai-capability-dictionary'
+      }
+    ]
+  },
+  {
+    text: 'IX. Engineering Excellence',
+    collapsed: false,
+    items: [
+      {
+        text: 'Code Quality',
+        link: '/zh-cn/appendix/9-engineering-excellence/code-quality-refactoring'
+      },
+      {
+        text: 'Testing Strategies',
+        link: '/zh-cn/appendix/9-engineering-excellence/testing-strategies'
+      },
+      {
+        text: 'Design Patterns',
+        link: '/zh-cn/appendix/9-engineering-excellence/design-patterns'
+      },
+      {
+        text: 'Security Thinking',
+        link: '/zh-cn/appendix/9-engineering-excellence/security-thinking'
+      },
+      {
+        text: 'Technical Writing',
+        link: '/zh-cn/appendix/9-engineering-excellence/technical-writing'
+      },
+      {
+        text: 'Open Source',
+        link: '/zh-cn/appendix/9-engineering-excellence/open-source-collaboration'
+      },
+      {
+        text: 'Tech Selection',
+        link: '/zh-cn/appendix/9-engineering-excellence/technology-selection'
+      }
+    ]
+  }
+]
+
+export const vibeStoriesSidebar = [
+  {
+    text: 'Vibe 故事',
+    collapsed: false,
+    items: [
+      {
+        text: '放弃月入过万，他在农村小学带孩子们"用AI赶苍蝇"',
+        link: '/zh-cn/vibe-stories/story-1'
+      },
+      {
+        text: '期末考试周，我偷偷用AI造了个"校园闲鱼"',
+        link: '/zh-cn/vibe-stories/story-2'
+      },
+      {
+        text: '我给每个学生，做了一个不会累的"学霸同桌"',
+        link: '/zh-cn/vibe-stories/story-3'
+      },
+      {
+        text: '48岁货车司机，熬了几个通宵，硬是用AI磕出一个出海工具站',
+        link: '/zh-cn/vibe-stories/story-4'
+      },
+      {
+        text: '七款 AI 编程工具对比',
+        link: '/zh-cn/stage-1/appendix-articles/example0-1/vibe-coding-tools-snake-game-tutorial'
+      },
+      {
+        text: '用设计和编程 Agent 构建网站',
+        link: '/zh-cn/stage-1/appendix-articles/example0-2/vibe-coding-tools-build-website-with-ai-coding-and-design-agents'
+      }
+    ]
+  }
+]
+
+export const vibeStoriesLabels = {
+  en: {
+    groupText: 'Vibe Stories',
+    items: [
+      'He Left a Five-Figure Monthly Salary to Help Rural School Kids "Use AI to Block Flies"',
+      'During Finals Week, I Secretly Built a "Campus Xianyu" with AI',
+      'I Built Each Student a Tireless "Straight-A Study Buddy"',
+      'At 48, a Truck Driver Pulled Several All-Nighters and Used AI to Build an Overseas Tool Site',
+      'Comparison of Seven AI Programming Tools',
+      'Building Websites with Design and Programming Agents'
+    ]
+  },
+  'ja-jp': {
+    groupText: 'Vibe ストーリー',
+    items: [
+      '月収10万超えを捨て、田舎の小学校で子供たちと「AIでハエを追い払う」',
+      '期末試験期間中、こっそりAIで「キャンパスメルカリ」を作った',
+      'すべての生徒に、疲れ知らずの「秀才の隣人」を作ってあげた',
+      '48歳のトラック運転手、数夜の徹夜でAIを使って海外向けツールサイトを作り上げた',
+      '7つのAIプログラミングツール比較',
+      'デザインAgentとプログラミングAgentでウェブサイトを構築'
+    ]
+  },
+  'zh-tw': {
+    groupText: 'Vibe 故事',
+    items: [
+      '放棄月入過萬，他在農村小學帶孩子們"用AI趕蒼蠅"',
+      '期末考試週，我偷偷用AI造了個"校園閒魚"',
+      '我給每個學生，做了一個不會累的"學霸同桌"',
+      '48歲貨車司機，熬了幾個通宵，硬是用AI磕出一個出海工具站',
+      '七款 AI 程式設計工具對比',
+      '用設計和程式設計 Agent 構建網站'
+    ]
+  },
+  'ko-kr': {
+    groupText: 'Vibe 스토리',
+    items: [
+      '월수입 100만원 이상을 포기하고 시골 초등학교에서 아이들과 "AI로 파리 쫓기"',
+      '기말고사 기간, 몰래 AI로 "캠퍼스 중고나라"를 만들었다',
+      '모든 학생에게 지치지 않는 "모범생 짝꿍"을 만들어주었다',
+      '48세 트럭 운전사, 며칠 밤을 새워 AI로 해외 툴 사이트를 만들었다',
+      '7가지 AI 프로그래밍 도구 비교',
+      '디자인 Agent와 코딩 Agent로 웹사이트 구축하기'
+    ]
+  },
+  'es-es': {
+    groupText: 'Historias Vibe',
+    items: [
+      'Dejó un salario de cinco cifras para ayudar a niños de escuelas rurales a "usar IA para espantar moscas"',
+      'Durante la semana de exámenes finales, construí secretamente un "Xianyu del campus" con IA',
+      'Construí a cada estudiante un "compañero de estudios sobresaliente" incansable',
+      'A los 48, un camionero pasó varias noches en vela y usó IA para construir un sitio de herramientas para el extranjero',
+      'Comparación de siete herramientas de programación con IA',
+      'Construyendo sitios web con agentes de diseño y programación'
+    ]
+  },
+  'fr-fr': {
+    groupText: 'Histoires Vibe',
+    items: [
+      'Il a quitté un salaire à cinq chiffres pour aider des écoliers ruraux à "utiliser l\'IA pour chasser les mouches"',
+      'Pendant la semaine des examens, j\'ai secrètement construit un "Xianyu du campus" avec l\'IA',
+      "J'ai construit à chaque étudiant un « camarade d'études brillant » infatigable",
+      "À 48 ans, un camionneur a passé plusieurs nuits blanches et a utilisé l'IA pour créer un site d'outils à l'international",
+      'Comparaison de sept outils de programmation IA',
+      'Construire des sites web avec des agents de design et de programmation'
+    ]
+  },
+  'de-de': {
+    groupText: 'Vibe-Geschichten',
+    items: [
+      'Er gab ein fünfstelliges Monatsgehalt auf, um Landschulkindern zu helfen, „mit KI Fliegen zu vertreiben"',
+      'Während der Abschlussprüfungswoche baute ich heimlich ein „Campus-Xianyu" mit KI',
+      'Ich baute jedem Schüler einen unermüdlichen „Einser-Schulkameraden"',
+      'Mit 48 zog ein LKW-Fahrer mehrere Nachtschichten durch und baute mit KI eine Übersee-Tool-Site',
+      'Vergleich von sieben KI-Programmierwerkzeugen',
+      'Erstellen von Websites mit Design- und Programmier-Agenten'
+    ]
+  },
+  'ar-sa': {
+    groupText: 'قصص Vibe',
+    items: [
+      'تخلى عن راتب بخمسة أرقام لمساعدة أطفال المدارس الريفية على "استخدام الذكاء الاصطناعي لطرد الذباب"',
+      'خلال أسبوع الامتحانات النهائية، بنيت سراً "شيان يو للحرم الجامعي" بالذكاء الاصطناعي',
+      'بنيت لكل طالب "رفيق دراسة متفوق" لا يكل',
+      'في سن 48، سائق شاحنة سهر عدة ليالي واستخدم الذكاء الاصطناعي لبناء موقع أدوات للخارج',
+      'مقارنة سبع أدوات برمجة بالذكاء الاصطناعي',
+      'بناء المواقع باستخدام وكلاء التصميم والبرمجة'
+    ]
+  },
+  'vi-vn': {
+    groupText: 'Câu chuyện Vibe',
+    items: [
+      'Từ bỏ thu nhập năm con số, anh ấy ở trường tiểu học nông thôn giúp trẻ em "dùng AI đuổi ruồi"',
+      'Tuần thi cuối kỳ, tôi bí mật dùng AI tạo ra "Campus Xianyu"',
+      'Tôi đã làm cho mỗi học sinh một "bạn cùng bàn học giỏi" không biết mệt',
+      'Ở tuổi 48, một tài xế xe tải đã thức nhiều đêm và dùng AI xây dựng một trang web công cụ ra nước ngoài',
+      'So sánh bảy công cụ lập trình AI',
+      'Xây dựng website bằng Agent thiết kế và lập trình'
+    ]
+  }
+}
+
+export const productManagerSidebar = [
+  {
+    text: '第一章 编程入门',
+    collapsed: false,
+    items: [
+      { text: '如何学习本课程', link: '/zh-cn/stage-1/learning-map/' },
+      {
+        text: 'AI 时代的编程初体验',
+        link: '/zh-cn/stage-1/ai-capabilities-through-games/'
+      }
+    ]
+  },
+  {
+    text: '第二章 开发环境与工具',
+    collapsed: false,
+    items: [
+      {
+        text: 'AI 编程工具介绍与使用',
+        link: '/zh-cn/stage-1/introduction-to-ai-ide/'
+      },
+      {
+        text: '从截图复刻：第一次模仿练习',
+        link: '/zh-cn/stage-1/clone-your-favorite-app/'
+      }
+    ]
+  },
+  {
+    text: '第三章 从问题到方案',
+    collapsed: false,
+    items: [
+      {
+        text: '1. 寻找真实问题',
+        link: '/zh-cn/stage-1/appendix-idea-sources/'
+      },
+      {
+        text: '2. 筛选值得做的方向',
+        link: '/zh-cn/stage-1/finding-great-idea/'
+      },
+      {
+        text: '3. 看懂真实需求（JTBD）',
+        link: '/zh-cn/stage-1/appendix-jobs-to-be-done/'
+      },
+      {
+        text: '4. 用访谈验证需求',
+        link: '/zh-cn/stage-1/appendix-mom-test/'
+      },
+      {
+        text: '5. 收敛方案（双钻模型）',
+        link: '/zh-cn/stage-1/appendix-double-diamond/'
+      }
+    ]
+  },
+  {
+    text: '第四章 从方案到作品',
+    collapsed: false,
+    items: [
+      {
+        text: '6. 做出可交互原型',
+        link: '/zh-cn/stage-1/building-prototype/'
+      },
+      {
+        text: '7. 接入真实 AI 能力',
+        link: '/zh-cn/stage-1/integrating-ai-capabilities/'
+      },
+      {
+        text: '8. 打磨并交付完整作品',
+        link: '/zh-cn/stage-1/complete-project-practice/'
+      }
+    ]
+  },
+  {
+    text: '附录',
+    collapsed: false,
+    items: [
+      {
+        text: '在魔搭社区发布你的 Vibe Coding 产物',
+        link: '/zh-cn/stage-1/appendix-modelscope-static-site/'
+      },
+      {
+        text: '产品思维基础',
+        link: '/zh-cn/stage-1/appendix-a-product-thinking/'
+      },
+      {
+        text: '从真实工作流里找 AI 场景',
+        link: '/zh-cn/stage-1/appendix-industry-scenarios/'
+      },
+      {
+        text: '常见问题与排错',
+        link: '/zh-cn/stage-1/appendix-b-common-errors/'
+      }
+    ]
+  }
+]
+
+export const appendixGroupLabels = {
+  en: [
+    'I. Computer Fundamentals',
+    'II. Tools & Environment',
+    'III. Browser & Frontend',
+    'IV. Server & Backend',
+    'V. Data',
+    'VI. Architecture',
+    'VII. Infrastructure',
+    'VIII. Artificial Intelligence',
+    'IX. Engineering Excellence'
+  ],
+  'ja-jp': [
+    'I. コンピュータ基礎',
+    'II. ツールと環境',
+    'III. ブラウザとフロントエンド',
+    'IV. サーバーとバックエンド',
+    'V. データ',
+    'VI. アーキテクチャ',
+    'VII. インフラ',
+    'VIII. 人工知能',
+    'IX. エンジニアリング品質'
+  ],
+  'zh-tw': [
+    'I. 計算機基礎',
+    'II. 工具與環境',
+    'III. 瀏覽器與前端',
+    'IV. 伺服器與後端',
+    'V. 資料',
+    'VI. 架構',
+    'VII. 基礎設施',
+    'VIII. 人工智慧',
+    'IX. 工程品質'
+  ],
+  'ko-kr': [
+    'I. 컴퓨터 기초',
+    'II. 도구와 환경',
+    'III. 브라우저와 프론트엔드',
+    'IV. 서버와 백엔드',
+    'V. 데이터',
+    'VI. 아키텍처',
+    'VII. 인프라',
+    'VIII. 인공지능',
+    'IX. 엔지니어링 품질'
+  ],
+  'es-es': [
+    'I. Fundamentos de Computación',
+    'II. Herramientas y Entorno',
+    'III. Navegador y Frontend',
+    'IV. Servidor y Backend',
+    'V. Datos',
+    'VI. Arquitectura',
+    'VII. Infraestructura',
+    'VIII. Inteligencia Artificial',
+    'IX. Excelencia de Ingeniería'
+  ],
+  'fr-fr': [
+    'I. Fondamentaux Informatique',
+    'II. Outils et Environnement',
+    'III. Navigateur et Frontend',
+    'IV. Serveur et Backend',
+    'V. Données',
+    'VI. Architecture',
+    'VII. Infrastructure',
+    'VIII. Intelligence Artificielle',
+    "IX. Excellence d'Ingénierie"
+  ],
+  'de-de': [
+    'I. Computergrundlagen',
+    'II. Werkzeuge und Umgebung',
+    'III. Browser und Frontend',
+    'IV. Server und Backend',
+    'V. Daten',
+    'VI. Architektur',
+    'VII. Infrastruktur',
+    'VIII. Künstliche Intelligenz',
+    'IX. Engineering-Qualität'
+  ],
+  'ar-sa': [
+    'I. أساسيات الحاسوب',
+    'II. الأدوات والبيئة',
+    'III. المتصفح والواجهة الأمامية',
+    'IV. الخادم والواجهة الخلفية',
+    'V. البيانات',
+    'VI. المعمارية',
+    'VII. البنية التحتية',
+    'VIII. الذكاء الاصطناعي',
+    'IX. جودة الهندسة'
+  ],
+  'vi-vn': [
+    'I. Nền tảng Máy tính',
+    'II. Công cụ và Môi trường',
+    'III. Trinh duyet va Frontend',
+    'IV. Máy chủ và Backend',
+    'V. Dữ liệu',
+    'VI. Kiến trúc',
+    'VII. Hạ tầng',
+    'VIII. Trí tuệ Nhân tạo',
+    'IX. Chất lượng Kỹ thuật'
+  ]
+}
+
+export const stage1SidebarLabels = {
+  'ja-jp': [
+    {
+      text: '第1章 プログラミング入門',
+      items: ['このコースの学び方', 'AI時代のプログラミング初体験']
+    },
+    {
+      text: '第2章 開発環境とツール',
+      items: [
+        'AIプログラミングツールの紹介と使い方',
+        'スクリーンショットから複製：初めての模倣練習'
+      ]
+    },
+    {
+      text: '第3章 問題から解決策へ',
+      items: [
+        '1. 本当の問題を見つける',
+        '2. 作る価値のある方向を選ぶ',
+        '3. 本当のニーズを理解する（JTBD）',
+        '4. インタビューでニーズを検証する',
+        '5. 解決策を絞り込む（ダブルダイヤモンド）'
+      ]
+    },
+    {
+      text: '第4章 解決策から作品へ',
+      items: [
+        '6. 操作できるプロトタイプを作る',
+        '7. 実際のAI機能を接続する',
+        '8. 磨き上げて完成作品を提出する'
+      ]
+    },
+    {
+      text: '付録',
+      items: [
+        'Vibe Codingの成果物をModelScopeで公開する',
+        'プロダクト思考の基礎',
+        '実際の業務フローからAI活用場面を探す',
+        'よくある問題とトラブルシューティング'
+      ]
+    }
+  ],
+  'zh-tw': [
+    {
+      text: '第一章 程式設計入門',
+      items: ['如何學習本課程', 'AI 時代的程式設計初體驗']
+    },
+    {
+      text: '第二章 開發環境與工具',
+      items: ['AI 程式設計工具介紹與使用', '從截圖复刻：第一次模仿練習']
+    },
+    {
+      text: '第三章 從問題到方案',
+      items: [
+        '1. 尋找真實問題',
+        '2. 篩選值得做的方向',
+        '3. 看懂真實需求（JTBD）',
+        '4. 用訪談驗證需求',
+        '5. 收斂方案（雙鑽模型）'
+      ]
+    },
+    {
+      text: '第四章 從方案到作品',
+      items: [
+        '6. 做出可互動原型',
+        '7. 接入真實 AI 能力',
+        '8. 打磨並交付完整作品'
+      ]
+    },
+    {
+      text: '附錄',
+      items: [
+        '在魔搭社群發布你的 Vibe Coding 作品',
+        '產品思維基礎',
+        '從真實工作流程中找 AI 場景',
+        '常見問題與除錯'
+      ]
+    }
+  ],
+  'es-es': [
+    {
+      text: 'Capítulo 1: Introducción a la Programación',
+      items: [
+        'Cómo aprender este curso',
+        'Primera experiencia de programación en la era de la IA'
+      ]
+    },
+    {
+      text: 'Capítulo 2: Entorno de Desarrollo y Herramientas',
+      items: [
+        'Introducción y uso de herramientas de programación con IA',
+        'Clonar desde capturas de pantalla: primer ejercicio de imitación'
+      ]
+    },
+    {
+      text: 'Capítulo 3: Del problema a la solución',
+      items: [
+        '1. Encontrar problemas reales',
+        '2. Elegir una dirección que merezca la pena',
+        '3. Entender necesidades reales (JTBD)',
+        '4. Validar necesidades con entrevistas',
+        '5. Concretar la solución (Double Diamond)'
+      ]
+    },
+    {
+      text: 'Capítulo 4: De la solución al producto',
+      items: [
+        '6. Crear un prototipo interactivo',
+        '7. Conectar capacidades reales de IA',
+        '8. Pulir y entregar un producto completo'
+      ]
+    },
+    {
+      text: 'Apéndice',
+      items: [
+        'Publica tu proyecto de Vibe Coding en ModelScope',
+        'Fundamentos del pensamiento de producto',
+        'Encontrar escenarios de IA en flujos de trabajo reales',
+        'Problemas comunes y solución de errores'
+      ]
+    }
+  ],
+  'fr-fr': [
+    {
+      text: 'Chapitre 1 : Introduction à la Programmation',
+      items: [
+        'Comment apprendre ce cours',
+        "Première expérience de programmation à l'ère de l'IA"
+      ]
+    },
+    {
+      text: 'Chapitre 2 : Environnement et Outils de Développement',
+      items: [
+        'Présentation et utilisation des outils de programmation IA',
+        "Cloner depuis des captures d'écran : premier exercice d'imitation"
+      ]
+    },
+    {
+      text: 'Chapitre 3 : Du problème à la solution',
+      items: [
+        '1. Trouver des problèmes réels',
+        '2. Choisir une direction qui mérite d’être construite',
+        '3. Comprendre les besoins réels (JTBD)',
+        '4. Valider les besoins par des entretiens',
+        '5. Resserrer la solution (Double Diamond)'
+      ]
+    },
+    {
+      text: 'Chapitre 4 : De la solution au produit',
+      items: [
+        '6. Construire un prototype interactif',
+        '7. Connecter de véritables capacités IA',
+        '8. Finaliser et livrer un produit complet'
+      ]
+    },
+    {
+      text: 'Annexe',
+      items: [
+        'Publier votre projet Vibe Coding sur ModelScope',
+        'Fondamentaux de la pensée produit',
+        'Trouver des usages IA dans les flux de travail réels',
+        'Problèmes courants et dépannage'
+      ]
+    }
+  ],
+  'de-de': [
+    {
+      text: 'Kapitel 1: Einführung in die Programmierung',
+      items: [
+        'Wie man diesen Kurs lernt',
+        'Erste Programmiererfahrung im KI-Zeitalter'
+      ]
+    },
+    {
+      text: 'Kapitel 2: Entwicklungsumgebung und Werkzeuge',
+      items: [
+        'Einführung und Nutzung von KI-Programmierwerkzeugen',
+        'Klonen aus Screenshots: Erste Imitationsübung'
+      ]
+    },
+    {
+      text: 'Kapitel 3: Vom Problem zur Lösung',
+      items: [
+        '1. Echte Probleme finden',
+        '2. Eine lohnende Richtung auswählen',
+        '3. Echte Bedürfnisse verstehen (JTBD)',
+        '4. Bedürfnisse durch Interviews prüfen',
+        '5. Die Lösung eingrenzen (Double Diamond)'
+      ]
+    },
+    {
+      text: 'Kapitel 4: Von der Lösung zum Produkt',
+      items: [
+        '6. Einen interaktiven Prototyp bauen',
+        '7. Echte KI-Funktionen anbinden',
+        '8. Ein vollständiges Produkt ausarbeiten und abgeben'
+      ]
+    },
+    {
+      text: 'Anhang',
+      items: [
+        'Dein Vibe-Coding-Projekt auf ModelScope veröffentlichen',
+        'Grundlagen des Produktdenkens',
+        'KI-Szenarien in echten Arbeitsabläufen finden',
+        'Häufige Probleme und Fehlerbehebung'
+      ]
+    }
+  ],
+  'ar-sa': [
+    {
+      text: 'الفصل الأول: مقدمة في البرمجة',
+      items: ['كيف تتعلم هذه الدورة', 'أول تجربة برمجة في عصر الذكاء الاصطناعي']
+    },
+    {
+      text: 'الفصل الثاني: بيئة التطوير والأدوات',
+      items: [
+        'مقدمة واستخدام أدوات البرمجة بالذكاء الاصطناعي',
+        'الاستنساخ من لقطات الشاشة: أول تمرين محاكاة'
+      ]
+    },
+    {
+      text: 'الفصل الثالث: من المشكلة إلى الحل',
+      items: [
+        '1. العثور على مشكلات حقيقية',
+        '2. اختيار اتجاه يستحق البناء',
+        '3. فهم الاحتياجات الحقيقية (JTBD)',
+        '4. التحقق من الاحتياجات بالمقابلات',
+        '5. تضييق الحل (الماسة المزدوجة)'
+      ]
+    },
+    {
+      text: 'الفصل الرابع: من الحل إلى المنتج',
+      items: [
+        '6. بناء نموذج أولي تفاعلي',
+        '7. ربط قدرات ذكاء اصطناعي حقيقية',
+        '8. صقل منتج كامل وتسليمه'
+      ]
+    },
+    {
+      text: 'الملحق',
+      items: [
+        'نشر مشروع Vibe Coding على ModelScope',
+        'أساسيات التفكير المنتجي',
+        'العثور على سيناريوهات الذكاء الاصطناعي في سير العمل الحقيقي',
+        'المشاكل الشائعة واستكشاف الأخطاء وإصلاحها'
+      ]
+    }
+  ],
+  'vi-vn': [
+    {
+      text: 'Chương 1: Giới thiệu Lập trình',
+      items: [
+        'Cách học khóa học này',
+        'Trải nghiệm lập trình đầu tiên trong thời đại AI'
+      ]
+    },
+    {
+      text: 'Chương 2: Môi trường Phát triển và Công cụ',
+      items: [
+        'Giới thiệu và sử dụng công cụ lập trình AI',
+        'Sao chép từ ảnh chụp màn hình: bài tập bắt chước đầu tiên'
+      ]
+    },
+    {
+      text: 'Chương 3: Từ vấn đề đến giải pháp',
+      items: [
+        '1. Tìm vấn đề thật',
+        '2. Chọn hướng đáng để làm',
+        '3. Hiểu nhu cầu thật (JTBD)',
+        '4. Kiểm chứng nhu cầu bằng phỏng vấn',
+        '5. Thu hẹp giải pháp (Double Diamond)'
+      ]
+    },
+    {
+      text: 'Chương 4: Từ giải pháp đến sản phẩm',
+      items: [
+        '6. Làm nguyên mẫu có thể tương tác',
+        '7. Kết nối năng lực AI thật',
+        '8. Hoàn thiện và bàn giao sản phẩm đầy đủ'
+      ]
+    },
+    {
+      text: 'Phụ lục',
+      items: [
+        'Đăng sản phẩm Vibe Coding lên ModelScope',
+        'Nền tảng tư duy sản phẩm',
+        'Tìm kịch bản AI trong quy trình làm việc thật',
+        'Các vấn đề thường gặp và khắc phục lỗi'
+      ]
+    }
+  ]
+}
+
+export const stage2SidebarLabels = {
+  'ja-jp': [
+    {
+      text: 'フロントエンド開発',
+      items: [
+        'NanoBanana 素材生産',
+        'Figma & MasterGo 入門',
+        'UI デザイン仕様とマルチプロダクトUI',
+        'Agent Skills でインターフェース美化',
+        'デザインプロトタイプからプロジェクトコードへ',
+        'モダンコンポーネントライブラリとUIアップグレード'
+      ]
+    },
+    {
+      text: 'バックエンド開発',
+      items: [
+        'データベースと Supabase 入門',
+        '大規模モデル補助インターフェース開発',
+        'Git & GitHub 入門ガイド',
+        'ウェブサイトを公開する（簡単編）：ワンクリック PaaS デプロイ',
+        'ウェブサイトを公開する（上級編）：VPS サーバーを自分で用意して公開',
+        'CLI Coding Agent プログラミングアシスタント',
+        'Stripe 決済統合'
+      ]
+    },
+    {
+      text: 'AI 能力付録',
+      items: ['Dify 入門とナレッジベース統合']
+    },
+    {
+      text: '総合プロジェクト',
+      items: [
+        'ホグワーツ肖像画を作ろう',
+        'AI マーケティングコピー SaaS',
+        'オンライン試験・管理システム',
+        'モダン AI 画像生成 SaaS',
+        'Dify ライクなエージェントプラットフォーム',
+        'スマート旅行プランニング Agent プラットフォーム',
+        'Spring Boot 映画推薦システム',
+        '生鮮ECマイクロサービスシステム',
+        'Go 交通データ分析プラットフォーム'
+      ]
+    }
+  ],
+  'zh-tw': [
+    {
+      text: '前端開發',
+      items: [
+        'NanoBanana 素材生產',
+        'Figma 與 MasterGo 入門',
+        'UI 設計規範與多產品介面',
+        '結合 Agent Skills 美化介面',
+        '設計原型到專案程式碼',
+        '現代元件庫與介面升級'
+      ]
+    },
+    {
+      text: '後端開發',
+      items: [
+        '資料庫與 Supabase 入門',
+        '大模型輔助介面開發',
+        'Git 與 GitHub 入門指南',
+        '把網站發到網上（簡單方式）：Vercel/Zeabur/CloudBase 一鍵發佈',
+        '把網站發到網上（進階方式）：自己買臺 VPS 伺服器搭建發佈',
+        'CLI Coding Agent 程式設計助手',
+        'Stripe 支付整合'
+      ]
+    },
+    {
+      text: 'AI 能力附錄',
+      items: ['Dify 入門與知識庫整合']
+    },
+    {
+      text: '綜合專案',
+      items: [
+        '一起做霍格沃茨畫像',
+        'AI 行銷文案 SaaS',
+        '線上考試與管理系統',
+        '現代 AI 生圖 SaaS',
+        '類 Dify 智能體平台',
+        '智慧旅遊規劃 Agent 平台',
+        'Spring Boot 電影推薦系統',
+        '生鮮電商微服務系統',
+        'Go 交通資料分析平台'
+      ]
+    }
+  ],
+  'ko-kr': [
+    {
+      text: '프론트엔드 개발',
+      items: [
+        'NanoBanana 에셋 생산',
+        'Figma & MasterGo 입문',
+        'UI 디자인 가이드라인과 멀티 프로덕트 UI',
+        'Agent Skills 인터페이스 미화',
+        '디자인 프로토타입에서 프로젝트 코드로',
+        '모던 컴포넌트 라이브러리와 UI 업그레이드'
+      ]
+    },
+    {
+      text: '백엔드 개발',
+      items: [
+        '데이터베이스와 Supabase 입문',
+        '대규모 모델 보조 인터페이스 개발',
+        'Git & GitHub 입문 가이드',
+        '웹사이트 공개하기 (쉬운 방법): 원클릭 PaaS 배포',
+        '웹사이트 공개하기 (고급): 직접 VPS 서버를 준비해서 배포하기',
+        'CLI Coding Agent 프로그래밍 어시스턴트',
+        'Stripe 결제 통합'
+      ]
+    },
+    {
+      text: 'AI 역량 부록',
+      items: ['Dify 입문과 지식 베이스 통합']
+    },
+    {
+      text: '종합 프로젝트',
+      items: [
+        '호그와트 초상화 만들기',
+        'AI 마케팅 카피 SaaS',
+        '온라인 시험 및 관리 시스템',
+        '모던 AI 이미지 생성 SaaS',
+        'Dify 유사 에이전트 플랫폼',
+        '스마트 여행 계획 Agent 플랫폼',
+        'Spring Boot 영화 추천 시스템',
+        '신선 식품 전자상거래 마이크로서비스',
+        'Go 교통 데이터 분석 플랫폼'
+      ]
+    }
+  ],
+  'es-es': [
+    {
+      text: 'Desarrollo Frontend',
+      items: [
+        'Producción de activos NanoBanana',
+        'Introducción a Figma y MasterGo',
+        'Especificaciones de diseño UI y multi-producto',
+        'Embellecimiento de interfaces con Agent Skills',
+        'De prototipo de diseño a código de proyecto',
+        'Bibliotecas de componentes modernas'
+      ]
+    },
+    {
+      text: 'Desarrollo Backend',
+      items: [
+        'Introducción a bases de datos y Supabase',
+        'Desarrollo de interfaces asistido por LLM',
+        'Guía de introducción a Git y GitHub',
+        'Publica tu sitio web (fácil): despliegue PaaS con un clic',
+        'Publica tu sitio web (avanzado): prepara tu propio VPS y despliega',
+        'CLI Coding Agent - Asistente de programación',
+        'Integración de pagos con Stripe'
+      ]
+    },
+    {
+      text: 'Apéndice de capacidades IA',
+      items: ['Introducción a Dify e integración de base de conocimientos']
+    },
+    {
+      text: 'Proyectos integrales',
+      items: [
+        'Creemos retratos de Hogwarts',
+        'SaaS de copywriting con IA',
+        'Sistema de exámenes en línea y gestión',
+        'SaaS moderno de generación de imágenes con IA',
+        'Plataforma de agentes tipo Dify',
+        'Plataforma de planificación de viajes con Agent',
+        'Sistema de recomendación de películas con Spring Boot',
+        'Sistema de microservicios de comercio electrónico de alimentos',
+        'Plataforma de análisis de datos de tráfico con Go'
+      ]
+    }
+  ],
+  'fr-fr': [
+    {
+      text: 'Développement Frontend',
+      items: [
+        'Production de ressources NanoBanana',
+        'Introduction à Figma et MasterGo',
+        'Spécifications UI et interfaces multi-produits',
+        'Embellissement des interfaces avec Agent Skills',
+        'Du prototype de design au code de projet',
+        'Bibliothèques de composants modernes'
+      ]
+    },
+    {
+      text: 'Développement Backend',
+      items: [
+        'Introduction aux bases de données et Supabase',
+        "Développement d'interfaces assisté par LLM",
+        "Guide d'introduction à Git et GitHub",
+        'Mettez votre site en ligne (facile) : déploiement PaaS en un clic',
+        'Mettez votre site en ligne (avancé) : préparez votre propre VPS',
+        'CLI Coding Agent - Assistant de programmation',
+        'Intégration de paiements avec Stripe'
+      ]
+    },
+    {
+      text: 'Annexe des capacités IA',
+      items: ['Introduction à Dify et intégration de base de connaissances']
+    },
+    {
+      text: 'Projets globaux',
+      items: [
+        'Créons des portraits de Poudlard',
+        'SaaS de rédaction IA',
+        "Système d'examens en ligne et de gestion",
+        "SaaS moderne de génération d'images IA",
+        "Plateforme d'agents de type Dify",
+        'Plateforme de planification de voyages avec Agent',
+        'Système de recommandation de films avec Spring Boot',
+        'Système de microservices e-commerce alimentaire',
+        "Plateforme d'analyse de données de trafic avec Go"
+      ]
+    }
+  ],
+  'de-de': [
+    {
+      text: 'Frontend-Entwicklung',
+      items: [
+        'NanoBanana Asset-Produktion',
+        'Einführung in Figma und MasterGo',
+        'UI-Design-Spezifikationen und Multi-Produkt-UI',
+        'Interface-Verschönerung mit Agent Skills',
+        'Vom Design-Prototyp zum Projektcode',
+        'Moderne Komponentenbibliotheken und UI-Upgrade'
+      ]
+    },
+    {
+      text: 'Backend-Entwicklung',
+      items: [
+        'Einführung in Datenbanken und Supabase',
+        'LLM-gestützte Schnittstellenentwicklung',
+        'Git und GitHub Einführungsleitfaden',
+        'Webseite online bringen (Einfach): One-Click-PaaS-Deployment',
+        'Webseite online bringen (Fortgeschritten): eigener VPS-Server',
+        'CLI Coding Agent Programmierassistent',
+        'Stripe-Zahlungsintegration'
+      ]
+    },
+    {
+      text: 'KI-Fähigkeiten Anhang',
+      items: ['Dify Einführung und Wissensdatenbank-Integration']
+    },
+    {
+      text: 'Übergreifende Projekte',
+      items: [
+        'Erstellen wir Hogwarts-Porträts',
+        'KI-Copywriting SaaS',
+        'Online-Prüfung und Managementsystem',
+        'Modernes KI-Bildgenerierungs-SaaS',
+        'Dify-ähnliche Agenten-Plattform',
+        'Intelligente Reiseplanungs-Agent-Plattform',
+        'Spring Boot Filmempfehlungssystem',
+        'Lebensmittel-E-Commerce-Microservices',
+        'Go Verkehrsanalyseplattform'
+      ]
+    }
+  ],
+  'ar-sa': [
+    {
+      text: 'تطوير الواجهة الأمامية',
+      items: [
+        'إنتاج أصول NanoBanana',
+        'مقدمة في Figma و MasterGo',
+        'مواصفات تصميم UI وواجهات متعددة المنتجات',
+        'تحسين الواجهات باستخدام Agent Skills',
+        'من النموذج الأولي إلى كود المشروع',
+        'مكتبات المكونات الحديثة وترقية الواجهة'
+      ]
+    },
+    {
+      text: 'تطوير الواجهة الخلفية',
+      items: [
+        'مقدمة في قواعد البيانات و Supabase',
+        'تطوير الواجهات بمساعدة النماذج الكبيرة',
+        'دليل مقدمة في Git و GitHub',
+        'انشر موقعك على الإنترنت (سهل): نشر PaaS بنقرة واحدة',
+        'انشر موقعك على الإنترنت (متقدم): جهّز خادم VPS خاصًا بك',
+        'مساعد برمجة CLI Coding Agent',
+        'تكامل مدفوعات Stripe'
+      ]
+    },
+    {
+      text: 'ملحق قدرات الذكاء الاصطناعي',
+      items: ['مقدمة في Dify وتكامل قاعدة المعرفة']
+    },
+    {
+      text: 'مشاريع شاملة',
+      items: [
+        'لنصنع صور هوغوورتس',
+        'SaaS كتابة التسويق بالذكاء الاصطناعي',
+        'نظام الامتحانات عبر الإنترنت والإدارة',
+        'SaaS حديث لتوليد الصور بالذكاء الاصطناعي',
+        'منصة وكلاء شبيهة بـ Dify',
+        'منصة تخطيط السفر الذكي مع Agent',
+        'نظام توصية الأفلام بـ Spring Boot',
+        'نظام الخدمات المصغرة للتجارة الإلكترونية للأغذية',
+        'منصة تحليل بيانات المرور بـ Go'
+      ]
+    }
+  ],
+  'vi-vn': [
+    {
+      text: 'Phát triển Frontend',
+      items: [
+        'Sản xuất tài sản NanoBanana',
+        'Giới thiệu Figma và MasterGo',
+        'Quy cách thiết kế UI và đa sản phẩm',
+        'Làm đẹp giao diện với Agent Skills',
+        'Từ nguyên mẫu thiết kế đến mã dự án',
+        'Thư viện thành phần hiện đại và nâng cấp UI'
+      ]
+    },
+    {
+      text: 'Phát triển Backend',
+      items: [
+        'Giới thiệu cơ sở dữ liệu và Supabase',
+        'Phát triển giao diện hỗ trợ bằng mô hình lớn',
+        'Hướng dẫn nhập môn Git và GitHub',
+        'Đưa website lên mạng (dễ dàng): triển khai PaaS một cú nhấp chuột',
+        'Đưa website lên mạng (nâng cao): tự chuẩn bị VPS để triển khai',
+        'Trợ lý lập trình CLI Coding Agent',
+        'Tích hợp thanh toán Stripe'
+      ]
+    },
+    {
+      text: 'Phụ lục năng lực AI',
+      items: ['Giới thiệu Dify và tích hợp cơ sở tri thức']
+    },
+    {
+      text: 'Dự án tổng hợp',
+      items: [
+        'Cùng làm chân dung Hogwarts',
+        'SaaS viết文案 AI',
+        'Hệ thống thi trực tuyến và quản lý',
+        'SaaS tạo ảnh AI hiện đại',
+        'Nền tảng Agent giống Dify',
+        'Nền tảng Agent lập kế hoạch du lịch thông minh',
+        'Hệ thống gợi ý phim Spring Boot',
+        'Hệ thống microservice thương mại điện tử thực phẩm',
+        'Nền tảng phân tích dữ liệu giao thông Go'
+      ]
+    }
+  ]
+}
+
+export const zhCnStage3Sidebar = [
+  {
+    text: 'Claude Code 深入浅出',
+    collapsed: false,
+    items: [
+      {
+        text: 'Claude Code 快速上手核心指南',
+        link: '/zh-cn/stage-3/core-skills/basics/'
+      },
+      {
+        text: 'Claude Code MCP 完全指南',
+        link: '/zh-cn/stage-3/core-skills/mcp/'
+      },
+      {
+        text: 'Claude Code Skills 完全指南',
+        link: '/zh-cn/stage-3/core-skills/skills/'
+      },
+      {
+        text: '如何让 Coding Tools 长时间工作',
+        link: '/zh-cn/stage-3/core-skills/long-running-tasks/'
+      },
+      {
+        text: 'Claude Agent Teams 完全指南',
+        link: '/zh-cn/stage-3/core-skills/agent-teams/'
+      },
+      {
+        text: 'Claude Code 工作流最佳实践',
+        link: '/zh-cn/stage-3/core-skills/workflow/'
+      },
+      {
+        text: 'Claude Code 手机远程开发',
+        link: '/zh-cn/stage-3/core-skills/mobile-development/'
+      },
+      {
+        text: 'Claude Agent SDK 完全指南',
+        link: '/zh-cn/stage-3/core-skills/claude-agent-sdk/'
+      }
+    ]
+  },
+  {
+    text: '规范驱动开发',
+    collapsed: false,
+    items: [
+      {
+        text: 'Claude Code Superpowers 工程级开发',
+        link: '/zh-cn/stage-3/core-skills/superpowers/'
+      },
+      {
+        text: '从 Vibe Coding 到 Spec Coding',
+        link: '/zh-cn/stage-3/core-skills/spec-coding/'
+      },
+      {
+        text: 'GitHub Issues 驱动开发实战',
+        link: '/zh-cn/stage-3/core-skills/github-iterative-development/'
+      }
+    ]
+  },
+  {
+    text: '多平台开发',
+    collapsed: false,
+    items: [
+      {
+        text: '从开发一个网页，到真正的程序，再到 AI 原生创作者',
+        link: '/zh-cn/stage-3/cross-platform/ai-native-creator/'
+      },
+      {
+        text: '如何选择你的应用该开发的平台',
+        link: '/zh-cn/stage-3/cross-platform/choose-platform/'
+      },
+      {
+        text: '如何构建微信小程序',
+        link: '/zh-cn/stage-3/cross-platform/wechat-miniprogram/'
+      },
+      {
+        text: '如何构建微信小程序（包含后端）',
+        link: '/zh-cn/stage-3/cross-platform/wechat-miniprogram-backend/'
+      },
+      {
+        text: '如何用 Jetpack Compose 开发 Android 原生应用',
+        link: '/zh-cn/stage-3/cross-platform/android-app/'
+      },
+      {
+        text: '如何用 SwiftUI 开发 iOS 原生应用',
+        link: '/zh-cn/stage-3/cross-platform/ios-app/'
+      },
+      {
+        text: '从零开始用 React Native + Expo 做一个门店巡检应用',
+        link: '/zh-cn/stage-3/cross-platform/react-native-expo/'
+      },
+      {
+        text: '如何开发 Flutter 跨平台应用',
+        link: '/zh-cn/stage-3/cross-platform/flutter-app/'
+      },
+      {
+        text: '如何开发 PWA 本地应用',
+        link: '/zh-cn/stage-3/cross-platform/pwa-local-app/'
+      },
+      {
+        text: '如何开发浏览器 AI 助手插件',
+        link: '/zh-cn/stage-3/cross-platform/browser-ai-extension/'
+      },
+      {
+        text: '如何开发跨平台 Electron 桌面程序',
+        link: '/zh-cn/stage-3/cross-platform/electron-voice-to-text/'
+      },
+      {
+        text: '如何用 Godot 开发横版、像素与 3D 游戏',
+        link: '/zh-cn/stage-3/cross-platform/godot-game-development/'
+      },
+      {
+        text: '如何在本地模拟链铸造 NFT',
+        link: '/zh-cn/stage-3/cross-platform/nft-minting/'
+      },
+      {
+        text: '如何开发 VS Code 插件',
+        link: '/zh-cn/stage-3/cross-platform/vscode-extension/'
+      },
+      {
+        text: '如何开发企业 Qt 设备客户端',
+        link: '/zh-cn/stage-3/cross-platform/qt-industrial-hmi/'
+      },
+      {
+        text: '如何构建属于自己的个人网页与学术博客',
+        link: '/zh-cn/stage-3/personal-brand/personal-website-blog/'
+      },
+      {
+        text: '如何把开发好的程序发布上架',
+        link: '/zh-cn/stage-3/cross-platform/app-publishing/'
+      }
+    ]
+  },
+  {
+    text: 'AI 能力强化',
+    collapsed: false,
+    items: [
+      {
+        text: '什么是 RAG 以及它如何工作',
+        link: '/zh-cn/stage-3/ai-advanced/rag-introduction/'
+      },
+      {
+        text: '中高级 RAG 与工作流编排 - 以 LangGraph 为例',
+        link: '/zh-cn/stage-3/ai-advanced/langgraph-advanced-rag/'
+      }
+    ]
+  }
+]
+
+export const stage3SidebarLabels = {
+  'ja-jp': [
+    {
+      text: 'Claude Code 深入解説',
+      items: [
+        'Claude Code クイックスタートガイド',
+        'MCP と Claude Code 完全ガイド',
+        'Claude Code Skills 完全ガイド',
+        'Coding Tools を長時間稼働させる方法',
+        'Claude Agent Teams 完全ガイド',
+        'Claude Code ワークフロー最佳实践',
+        'Claude Code モバイルリモート開発',
+        'Claude Agent SDK 完全ガイド'
+      ]
+    },
+    {
+      text: '仕様駆動開発',
+      items: [
+        'Claude Code Superpowers 実践開発',
+        'Vibe Coding から Spec Coding へ',
+        'GitHub Issues で AI 開発を最後まで進める'
+      ]
+    },
+    {
+      text: 'マルチプラットフォーム開発',
+      items: [
+        'Web ページから実用ソフトウェア、そして AI ネイティブ製品へ',
+        'アプリ開発プラットフォームの選び方',
+        'WeChat ミニプログラムの構築',
+        'WeChat ミニプログラム（バックエンド付き）',
+        'Android アプリ開発（Jetpack Compose）',
+        'iOS アプリ開発（SwiftUI）',
+        'React Native + Expo で店舗点検アプリを作る',
+        'Flutter クロスプラットフォームアプリの開発',
+        'PWA ローカルアプリの開発',
+        'ブラウザ AI アシスタント拡張機能の開発',
+        'クロスプラットフォーム Electron デスクトップアプリ',
+        'Godot で横スクロール・ピクセル・3D ゲームを作る',
+        'NFT の迅速開発とミント',
+        'VS Code 拡張機能の開発',
+        '産業用 Qt デスクトップアプリの開発',
+        '個人ウェブサイトと学術ブログの構築',
+        '完成したアプリを公開・ストア配信する'
+      ]
+    },
+    {
+      text: 'AI 能力強化',
+      items: [
+        'RAG とは何か、その仕組み',
+        '中級・上級 RAG とワークフロー編成 - LangGraph を例に'
+      ]
+    }
+  ],
+  'zh-tw': [
+    {
+      text: 'Claude Code 深入淺出',
+      items: [
+        'Claude Code 快速上手核心指南',
+        'Claude Code MCP 完全指南',
+        'Claude Code Skills 完全指南',
+        '如何讓 Coding Tools 長時間工作',
+        'Claude Agent Teams 完全指南',
+        'Claude Code 工作流最佳實踐',
+        'Claude Code 手機遠端開發',
+        'Claude Agent SDK 完全指南'
+      ]
+    },
+    {
+      text: '規範驅動開發',
+      items: [
+        'Claude Code Superpowers 工程級開發',
+        '從 Vibe Coding 到 Spec Coding',
+        '用 GitHub Issues 推進 AI 全流程開發'
+      ]
+    },
+    {
+      text: '多平台開發',
+      items: [
+        '從網頁到完整程式，再到 AI 原生產品',
+        '如何選擇應用開發平台',
+        '如何構建微信小程式',
+        '如何構建微信小程式（包含後端）',
+        '如何構建安卓程式 - Compose 原生開發',
+        '如何構建 iOS 程式 - SwiftUI 原生開發',
+        '用 React Native + Expo 建立門市巡檢 App',
+        '如何用 Flutter 開發跨平台 App',
+        '如何開發 PWA 本地應用',
+        '如何開發瀏覽器 AI 助手插件',
+        '如何開發跨平台 Electron 桌面程式',
+        '如何用 Godot 開發橫向、像素與 3D 遊戲',
+        '如何快速開發並鑄造 NFT',
+        '如何開發 VS Code 插件',
+        '如何開發工業級 Qt 桌面應用',
+        '如何構建個人網頁與學術部落格',
+        '如何把做好的程式發布上架'
+      ]
+    },
+    {
+      text: 'AI 能力強化',
+      items: [
+        '什麼是 RAG 以及它如何工作',
+        '中高級 RAG 與工作流編排 - 以 LangGraph 為例'
+      ]
+    }
+  ],
+  'ko-kr': [
+    {
+      text: 'Claude Code 심층 가이드',
+      items: [
+        'Claude Code 퀵스타트 핵심 가이드',
+        'MCP와 Claude Code 완전 가이드',
+        'Claude Code Skills 완전 가이드',
+        'Coding Tools를 장시간 작동시키는 방법',
+        'Claude Agent Teams 완전 가이드',
+        'Claude Code 워크플로우 모범 사례',
+        'Claude Code 모바일 원격 개발',
+        'Claude Agent SDK 완전 가이드'
+      ]
+    },
+    {
+      text: '명세 주도 개발',
+      items: [
+        'Claude Code Superpowers 엔지니어링 개발',
+        'Vibe Coding에서 Spec Coding으로',
+        'GitHub Issues로 AI 개발을 끝까지 진행하기'
+      ]
+    },
+    {
+      text: '멀티플랫폼 개발',
+      items: [
+        '웹페이지에서 실제 프로그램, AI 네이티브 제품으로',
+        '앱 개발 플랫폼 선택 방법',
+        'WeChat 미니프로그램 구축',
+        'WeChat 미니프로그램 (백엔드 포함)',
+        'Android 앱 개발 (Jetpack Compose)',
+        'iOS 앱 개발 (SwiftUI)',
+        'React Native + Expo로 매장 점검 앱 만들기',
+        'Flutter 크로스 플랫폼 앱 개발',
+        'PWA 로컬 앱 개발',
+        '브라우저 AI 어시스턴트 확장 프로그램 개발',
+        '크로스 플랫폼 Electron 데스크톱 앱',
+        'Godot로 횡스크롤·픽셀·3D 게임 만들기',
+        'NFT 빠른 개발 및 민팅',
+        'VS Code 확장 프로그램 개발',
+        '산업용 Qt 데스크톱 앱 개발',
+        '개인 웹사이트 및 학술 블로그 구축',
+        '완성한 앱을 배포하고 스토어에 출시하기'
+      ]
+    },
+    {
+      text: 'AI 역량 강화',
+      items: [
+        'RAG란 무엇이며 어떻게 작동하는가',
+        '중고급 RAG와 워크플로우 오케스트레이션 - LangGraph 예제'
+      ]
+    }
+  ],
+  'es-es': [
+    {
+      text: 'Claude Code en Profundidad',
+      items: [
+        'Guía Core de Inicio Rápido de Claude Code',
+        'Guía Completa de MCP y Claude Code',
+        'Guía Completa de Claude Code Skills',
+        'Cómo hacer que Coding Tools trabajen largas horas',
+        'Guía Completa de Claude Agent Teams',
+        'Mejores Prácticas de Flujo de Trabajo con Claude Code',
+        'Desarrollo Remoto Móvil con Claude Code',
+        'Guía Completa de Claude Agent SDK'
+      ]
+    },
+    {
+      text: 'Desarrollo guiado por especificaciones',
+      items: [
+        'Claude Code Superpowers para Desarrollo de Producción',
+        'De Vibe Coding a Spec Coding',
+        'Desarrollo integral con IA y GitHub Issues'
+      ]
+    },
+    {
+      text: 'Desarrollo Multiplataforma',
+      items: [
+        'De una página web a un programa real y un producto nativo de IA',
+        'Cómo elegir la plataforma para tu aplicación',
+        'Cómo construir un Mini Programa de WeChat',
+        'Cómo construir un Mini Programa de WeChat (con Backend)',
+        'Cómo construir una app Android (Jetpack Compose)',
+        'Cómo construir una app iOS (SwiftUI)',
+        'Crear una app de inspección con React Native y Expo',
+        'Desarrollar una aplicación multiplataforma con Flutter',
+        'Cómo desarrollar una aplicación PWA local',
+        'Cómo desarrollar una extensión de navegador AI',
+        'Cómo desarrollar una app de escritorio Electron',
+        'Crear juegos de plataformas, pixel art y 3D con Godot',
+        'Cómo desarrollar y acuñar NFTs rápidamente',
+        'Cómo desarrollar una extensión de VS Code',
+        'Cómo desarrollar una app Qt industrial',
+        'Cómo construir tu sitio web personal y blog académico',
+        'Cómo publicar la aplicación terminada'
+      ]
+    },
+    {
+      text: 'IA Avanzada',
+      items: [
+        'Qué es RAG y cómo funciona',
+        'RAG avanzado y orquestación de flujos con LangGraph'
+      ]
+    }
+  ],
+  'fr-fr': [
+    {
+      text: 'Claude Code en Profondeur',
+      items: [
+        'Guide Core de Démarrage Rapide Claude Code',
+        'Guide Complet MCP et Claude Code',
+        'Guide Complet Claude Code Skills',
+        'Comment faire travailler les Coding Tools sur de longues périodes',
+        'Guide Complet Claude Agent Teams',
+        'Meilleures Pratiques de Workflow Claude Code',
+        'Développement à Distance sur Mobile avec Claude Code',
+        'Guide Complet Claude Agent SDK'
+      ]
+    },
+    {
+      text: 'Développement piloté par les spécifications',
+      items: [
+        'Claude Code Superpowers pour le Développement Production',
+        'Du Vibe Coding au Spec Coding',
+        'Développer de bout en bout avec GitHub Issues'
+      ]
+    },
+    {
+      text: 'Développement Multiplateforme',
+      items: [
+        'De la page web au logiciel, puis au produit natif IA',
+        'Comment choisir la plateforme pour votre application',
+        'Comment construire un Mini Programme WeChat',
+        'Comment construire un Mini Programme WeChat (avec Backend)',
+        'Comment construire une app Android (Jetpack Compose)',
+        'Comment construire une app iOS (SwiftUI)',
+        'Créer une application d’inspection avec React Native et Expo',
+        'Développer une application multiplateforme avec Flutter',
+        'Comment développer une application PWA locale',
+        'Comment développer une extension de navigateur AI',
+        'Comment développer une app bureau Electron',
+        'Créer des jeux de plateforme, pixel art et 3D avec Godot',
+        'Comment développer et minter des NFTs rapidement',
+        'Comment développer une extension VS Code',
+        'Comment développer une app Qt industrielle',
+        'Comment construire votre site personnel et blog académique',
+        'Publier et distribuer l’application terminée'
+      ]
+    },
+    {
+      text: 'IA Avancée',
+      items: [
+        "Qu'est-ce que le RAG et comment ça fonctionne",
+        'RAG avancé et orchestration de workflows avec LangGraph'
+      ]
+    }
+  ],
+  'de-de': [
+    {
+      text: 'Claude Code Vertieft',
+      items: [
+        'Claude Code Schnellstart-Kernleitfaden',
+        'MCP und Claude Code Vollständiger Leitfaden',
+        'Claude Code Skills Vollständiger Leitfaden',
+        'Wie man Coding Tools über lange Zeiträume arbeiten lässt',
+        'Claude Agent Teams Vollständiger Leitfaden',
+        'Claude Code Workflow Best Practices',
+        'Claude Code Mobile Remote-Entwicklung',
+        'Claude Agent SDK Vollständiger Leitfaden'
+      ]
+    },
+    {
+      text: 'Spezifikationsgetriebene Entwicklung',
+      items: [
+        'Claude Code Superpowers für Produktionsentwicklung',
+        'Vom Vibe Coding zum Spec Coding',
+        'KI-Softwareentwicklung mit GitHub Issues'
+      ]
+    },
+    {
+      text: 'Plattformübergreifende Entwicklung',
+      items: [
+        'Von der Webseite zur Software und zum KI-nativen Produkt',
+        'Wie man die richtige Plattform für seine App wählt',
+        'Wie man ein WeChat Mini-Programm erstellt',
+        'Wie man ein WeChat Mini-Programm (mit Backend) erstellt',
+        'Wie man eine Android-App erstellt (Jetpack Compose)',
+        'Wie man eine iOS-App erstellt (SwiftUI)',
+        'Eine Filialprüfungs-App mit React Native und Expo bauen',
+        'Eine plattformübergreifende App mit Flutter entwickeln',
+        'Wie man eine lokale PWA-App entwickelt',
+        'Wie man eine Browser-KI-Assistenten-Erweiterung entwickelt',
+        'Wie man eine plattformübergreifende Electron-Desktop-App entwickelt',
+        'Plattform-, Pixel- und 3D-Spiele mit Godot bauen',
+        'Wie man schnell NFTs entwickelt und mintet',
+        'Wie man eine VS Code-Erweiterung entwickelt',
+        'Wie man eine industrielle Qt-Desktop-App entwickelt',
+        'Wie man eine persönliche Website und einen akademischen Blog erstellt',
+        'Die fertige Anwendung veröffentlichen'
+      ]
+    },
+    {
+      text: 'KI Fortgeschritten',
+      items: [
+        'Was ist RAG und wie funktioniert es',
+        'Fortgeschrittenes RAG und Workflow-Orchestrierung mit LangGraph'
+      ]
+    }
+  ],
+  'ar-sa': [
+    {
+      text: 'Claude Code بعمق',
+      items: [
+        'دليل البدء السريع لـ Claude Code',
+        'الدليل الشامل لـ MCP و Claude Code',
+        'الدليل الشامل لـ Claude Code Skills',
+        'كيف تجعل أدوات البرمجة تعمل لفترات طويلة',
+        'الدليل الشامل لـ Claude Agent Teams',
+        'أفضل ممارسات سير عمل Claude Code',
+        'التطوير عن بُعد عبر الهاتف مع Claude Code',
+        'الدليل الشامل لـ Claude Agent SDK'
+      ]
+    },
+    {
+      text: 'التطوير الموجّه بالمواصفات',
+      items: [
+        'Claude Code Superpowers للتطوير على مستوى الإنتاج',
+        'من Vibe Coding إلى Spec Coding',
+        'تطوير البرمجيات بالذكاء الاصطناعي عبر GitHub Issues'
+      ]
+    },
+    {
+      text: 'التطوير متعدد المنصات',
+      items: [
+        'من صفحة ويب إلى برنامج حقيقي ومنتج أصلي للذكاء الاصطناعي',
+        'كيفية اختيار المنصة المناسبة لتطبيقك',
+        'كيفية بناء برنامج WeChat المصغر',
+        'كيفية بناء برنامج WeChat المصغر (مع خلفية)',
+        'كيفية بناء تطبيق Android (Jetpack Compose)',
+        'كيفية بناء تطبيق iOS (SwiftUI)',
+        'إنشاء تطبيق تفتيش متجر باستخدام React Native وExpo',
+        'تطوير تطبيق متعدد المنصات باستخدام Flutter',
+        'كيفية تطوير تطبيق PWA محلي',
+        'كيفية تطوير إضافة مساعد ذكاء اصطناعي للمتصفح',
+        'كيفية تطوير تطبيق سطح مكتب Electron',
+        'إنشاء ألعاب منصات وPixel Art و3D باستخدام Godot',
+        'كيفية تطوير وسك NFTs بسرعة',
+        'كيفية تطوير إضافة VS Code',
+        'كيفية تطوير تطبيق Qt صناعي',
+        'كيفية بناء موقعك الشخصي ومدونتك الأكاديمية',
+        'نشر التطبيق المكتمل في المتاجر'
+      ]
+    },
+    {
+      text: 'الذكاء الاصطناعي المتقدم',
+      items: [
+        'ما هو RAG وكيف يعمل',
+        'RAG المتقدم وتنسيق سير العمل مع LangGraph'
+      ]
+    }
+  ],
+  'vi-vn': [
+    {
+      text: 'Claude Code Chuyên Sâu',
+      items: [
+        'Hướng dẫn Khởi động Nhanh Claude Code',
+        'Hướng dẫn Toàn diện MCP và Claude Code',
+        'Hướng dẫn Toàn diện Claude Code Skills',
+        'Cách làm cho Coding Tools hoạt động trong thời gian dài',
+        'Hướng dẫn Toàn diện Claude Agent Teams',
+        'Thực hành Tốt nhất Workflow Claude Code',
+        'Phát triển Từ xa trên Mobile với Claude Code',
+        'Hướng dẫn Toàn diện Claude Agent SDK'
+      ]
+    },
+    {
+      text: 'Phát triển theo đặc tả',
+      items: [
+        'Claude Code Superpowers cho Phát triển Cấp Production',
+        'Từ Vibe Coding đến Spec Coding',
+        'Phát triển phần mềm trọn quy trình với GitHub Issues'
+      ]
+    },
+    {
+      text: 'Phát triển Đa nền tảng',
+      items: [
+        'Từ trang web đến phần mềm và sản phẩm AI-native',
+        'Cách chọn nền tảng phù hợp cho ứng dụng của bạn',
+        'Cách xây dựng WeChat Mini Program',
+        'Cách xây dựng WeChat Mini Program (có Backend)',
+        'Cách xây dựng ứng dụng Android (Jetpack Compose)',
+        'Cách xây dựng ứng dụng iOS (SwiftUI)',
+        'Xây dựng ứng dụng kiểm tra bằng React Native và Expo',
+        'Phát triển ứng dụng đa nền tảng với Flutter',
+        'Cách phát triển ứng dụng PWA cục bộ',
+        'Cách phát triển tiện ích trợ lý AI cho trình duyệt',
+        'Cách phát triển ứng dụng desktop Electron đa nền tảng',
+        'Làm game platform, pixel art và 3D bằng Godot',
+        'Cách phát triển và đúc NFT nhanh chóng',
+        'Cách phát triển tiện ích mở rộng VS Code',
+        'Cách phát triển ứng dụng desktop Qt công nghiệp',
+        'Cách xây dựng trang web cá nhân và blog học thuật',
+        'Phát hành ứng dụng đã hoàn thành'
+      ]
+    },
+    {
+      text: 'AI Nâng cao',
+      items: [
+        'RAG là gì và cách nó hoạt động',
+        'RAG nâng cao và điều phối workflow với LangGraph'
+      ]
+    }
+  ]
+}
+
+export const zhCnSidebar = {
+  '/zh-cn/vibe-stories/': vibeStoriesSidebar,
+  '/zh-cn/stage-1/appendix-articles/example0-1/': vibeStoriesSidebar,
+  '/zh-cn/stage-1/appendix-articles/example0-2/': vibeStoriesSidebar,
+  '/zh-cn/stage-1/': productManagerSidebar,
+  '/zh-cn/stage-2/': zhCnStage2Sidebar,
+  '/zh-cn/stage-3/': zhCnStage3Sidebar,
+  '/zh-cn/guide/': [
+    {
+      text: '课程指南',
+      items: [{ text: '课程介绍', link: '/zh-cn/guide/introduction' }]
+    }
+  ],
+  '/zh-cn/extra/': [
+    {
+      text: 'Extra 扩展知识（旧版，已迁移到 Stage 2/3）',
+      items: [
+        {
+          text: 'Extra 1: Git & GitHub',
+          link: '/zh-cn/stage-2/backend/git-workflow/'
+        },
+        {
+          text: 'Extra 2: What is API',
+          link: '/zh-cn/stage-2/backend/ai-interface-code/'
+        },
+        {
+          text: 'Extra 5: What is RAG',
+          link: '/zh-cn/stage-3/ai-advanced/rag-introduction/'
+        },
+        {
+          text: 'Extra 6: Zeabur Deployment',
+          link: '/zh-cn/stage-2/backend/zeabur-deployment/'
+        },
+        {
+          text: 'Extra 7: CLI AI Tools & TDD',
+          link: '/zh-cn/stage-2/backend/modern-cli/'
+        }
+      ]
+    }
+  ],
+  '/zh-cn/examples/': [
+    {
+      text: 'Examples 实战案例（旧版，已迁移到 Stage 0/3）',
+      items: [
+        {
+          text: 'Ex 0.1: Snake Game',
+          link: '/zh-cn/stage-1/appendix-articles/example0-1/vibe-coding-tools-snake-game-tutorial'
+        },
+        {
+          text: 'Ex 0.2: Build Website with AI',
+          link: '/zh-cn/stage-1/appendix-articles/example0-2/vibe-coding-tools-build-website-with-ai-coding-and-design-agents'
+        }
+      ]
+    }
+  ],
+  '/zh-cn/project/': [
+    {
+      text: 'Project 文档（旧版，已迁移到 Stage 2）',
+      items: [
+        {
+          text: '一起做霍格沃茨画像',
+          link: '/zh-cn/stage-2/frontend/hogwarts-portraits/'
+        },
+        {
+          text: 'Supabase 数据库',
+          link: '/zh-cn/stage-2/backend/database-supabase/'
+        },
+        {
+          text: 'Dify & Knowledge Base',
+          link: '/zh-cn/stage-2/ai-capabilities/dify-knowledge-base/'
+        }
+      ]
+    }
+  ],
+  '/zh-cn/appendix/': [
+    {
+      text: '一、计算机是怎么回事',
+      collapsed: false,
+      items: [
+        {
+          text: '全栈开发导论：Vibe Coding 时代的计算机知识地图',
+          link: '/zh-cn/appendix/1-computer-fundamentals/vibe-coding-fullstack'
+        },
+        {
+          text: '计算机系统全景：从开机到访问网站',
+          link: '/zh-cn/appendix/1-computer-fundamentals/power-on-to-web'
+        },
+        {
+          text: '数字电路基础：从晶体管到 CPU',
+          link: '/zh-cn/appendix/1-computer-fundamentals/transistor-to-cpu'
+        },
+        {
+          text: '计算机组成原理',
+          link: '/zh-cn/appendix/1-computer-fundamentals/computer-organization'
+        },
+        {
+          text: '操作系统原理：进程、内存与文件系统',
+          link: '/zh-cn/appendix/1-computer-fundamentals/operating-systems'
+        },
+        {
+          text: '数据表示原理：编码、存储与传输',
+          link: '/zh-cn/appendix/1-computer-fundamentals/data-encoding-storage'
+        },
+        {
+          text: '计算机网络：从输入网址到渲染页面',
+          link: '/zh-cn/appendix/1-computer-fundamentals/computer-networks'
+        },
+        {
+          text: '数据结构导论',
+          link: '/zh-cn/appendix/1-computer-fundamentals/data-structures'
+        },
+        {
+          text: '算法导论',
+          link: '/zh-cn/appendix/1-computer-fundamentals/algorithm-thinking'
+        },
+        {
+          text: '编程语言概念：范式、演化与选型',
+          link: '/zh-cn/appendix/1-computer-fundamentals/programming-languages'
+        },
+        {
+          text: '编译原理',
+          link: '/zh-cn/appendix/1-computer-fundamentals/compilers'
+        },
+        {
+          text: '类型系统导论',
+          link: '/zh-cn/appendix/1-computer-fundamentals/type-systems'
+        }
+      ]
+    },
+    {
+      text: '二、开发环境与工具',
+      collapsed: false,
+      items: [
+        {
+          text: '集成开发环境 (IDE) 基础',
+          link: '/zh-cn/appendix/2-development-tools/ide-basics'
+        },
+        {
+          text: '命令行与 Shell 脚本',
+          link: '/zh-cn/appendix/2-development-tools/command-line-shell'
+        },
+        {
+          text: 'Git：代码的时光机',
+          link: '/zh-cn/appendix/2-development-tools/git-version-control'
+        },
+        {
+          text: '环境变量与 PATH',
+          link: '/zh-cn/appendix/2-development-tools/environment-path'
+        },
+        {
+          text: '端口与 localhost',
+          link: '/zh-cn/appendix/2-development-tools/ports-localhost'
+        },
+        {
+          text: 'SSH 与密钥认证',
+          link: '/zh-cn/appendix/2-development-tools/ssh-authentication'
+        },
+        {
+          text: '包管理器（npm / pip / cargo）',
+          link: '/zh-cn/appendix/2-development-tools/package-managers'
+        },
+        {
+          text: '调试的艺术',
+          link: '/zh-cn/appendix/2-development-tools/debugging-art'
+        },
+        {
+          text: '正则表达式',
+          link: '/zh-cn/appendix/2-development-tools/regex'
+        }
+      ]
+    },
+    {
+      text: '三、浏览器与前端',
+      collapsed: false,
+      items: [
+        {
+          text: 'JavaScript 语言深入',
+          link: '/zh-cn/appendix/3-browser-and-frontend/javascript-deep-dive'
+        },
+        {
+          text: 'TypeScript：给 JS 加上类型系统',
+          link: '/zh-cn/appendix/3-browser-and-frontend/typescript'
+        },
+        {
+          text: '前端框架对比（React / Vue / Svelte / Angular）',
+          link: '/zh-cn/appendix/3-browser-and-frontend/frontend-frameworks'
+        },
+        {
+          text: '浏览器渲染管道',
+          link: '/zh-cn/appendix/3-browser-and-frontend/browser-as-os-rendering'
+        },
+        {
+          text: 'HTML / CSS 布局体系',
+          link: '/zh-cn/appendix/3-browser-and-frontend/html-css-layout'
+        },
+        {
+          text: 'JavaScript 运行时',
+          link: '/zh-cn/appendix/3-browser-and-frontend/javascript-runtime'
+        },
+        {
+          text: '前端框架的本质',
+          link: '/zh-cn/appendix/3-browser-and-frontend/frontend-framework-nature'
+        },
+        {
+          text: '状态管理哲学',
+          link: '/zh-cn/appendix/3-browser-and-frontend/state-management'
+        },
+        {
+          text: '路由与导航',
+          link: '/zh-cn/appendix/3-browser-and-frontend/routing-navigation'
+        },
+        {
+          text: '图形与动画（Canvas / SVG / WebGL）',
+          link: '/zh-cn/appendix/3-browser-and-frontend/graphics-animation'
+        },
+        {
+          text: '实时通信（WebSocket / SSE）',
+          link: '/zh-cn/appendix/3-browser-and-frontend/realtime-communication'
+        },
+        {
+          text: '网页性能的度量与优化',
+          link: '/zh-cn/appendix/3-browser-and-frontend/web-performance'
+        },
+        {
+          text: '前端工程化全貌',
+          link: '/zh-cn/appendix/3-browser-and-frontend/frontend-engineering'
+        },
+        {
+          text: '前端项目文件目录设计',
+          link: '/zh-cn/appendix/3-browser-and-frontend/frontend-project-architecture'
+        },
+        {
+          text: '无障碍与国际化',
+          link: '/zh-cn/appendix/3-browser-and-frontend/a11n-i18n'
+        }
+      ]
+    },
+    {
+      text: '四、服务器与后端',
+      collapsed: false,
+      items: [
+        {
+          text: '后端语言对比（Node.js / Go / Java / Rust）',
+          link: '/zh-cn/appendix/4-server-and-backend/backend-languages'
+        },
+        {
+          text: '客户端语言对比（Swift / Kotlin / Dart）',
+          link: '/zh-cn/appendix/4-server-and-backend/client-languages'
+        },
+        {
+          text: '跨平台方案对比（React Native / Flutter / Electron / Tauri）',
+          link: '/zh-cn/appendix/4-server-and-backend/cross-platform'
+        },
+        {
+          text: 'HTTP 协议',
+          link: '/zh-cn/appendix/4-server-and-backend/http-protocol'
+        },
+        {
+          text: '一个请求的完整旅程',
+          link: '/zh-cn/appendix/4-server-and-backend/request-journey'
+        },
+        {
+          text: 'Web 框架的本质',
+          link: '/zh-cn/appendix/4-server-and-backend/web-frameworks'
+        },
+        {
+          text: 'API 入门',
+          link: '/zh-cn/appendix/4-server-and-backend/api-intro'
+        },
+        {
+          text: 'API 设计哲学（REST / GraphQL / gRPC）',
+          link: '/zh-cn/appendix/4-server-and-backend/api-design'
+        },
+        {
+          text: '序列化与数据格式',
+          link: '/zh-cn/appendix/4-server-and-backend/serialization'
+        },
+        {
+          text: '认证与授权体系',
+          link: '/zh-cn/appendix/4-server-and-backend/auth-authorization'
+        },
+        {
+          text: '并发、异步与多线程',
+          link: '/zh-cn/appendix/4-server-and-backend/concurrency-async'
+        },
+        {
+          text: '缓存的层次与策略',
+          link: '/zh-cn/appendix/4-server-and-backend/caching'
+        },
+        {
+          text: '消息队列与事件驱动',
+          link: '/zh-cn/appendix/4-server-and-backend/message-queues'
+        },
+        {
+          text: '异步任务队列与生产消费模型',
+          link: '/zh-cn/appendix/4-server-and-backend/async-task-queues'
+        },
+        {
+          text: '限流与背压控制',
+          link: '/zh-cn/appendix/4-server-and-backend/rate-limiting-backpressure'
+        },
+        {
+          text: '搜索引擎原理',
+          link: '/zh-cn/appendix/4-server-and-backend/search-engines'
+        },
+        {
+          text: '文件存储与对象存储',
+          link: '/zh-cn/appendix/4-server-and-backend/file-storage'
+        },
+        {
+          text: '后端分层架构',
+          link: '/zh-cn/appendix/4-server-and-backend/backend-layered-architecture'
+        },
+        {
+          text: '后端项目文件目录设计',
+          link: '/zh-cn/appendix/4-server-and-backend/backend-project-architecture'
+        },
+        {
+          text: '领域特定语言（DSL）',
+          link: '/zh-cn/appendix/4-server-and-backend/domain-specific-languages'
+        }
+      ]
+    },
+    {
+      text: '五、数据',
+      collapsed: false,
+      items: [
+        {
+          text: '数据库原理（索引 / 事务 / 查询优化）',
+          link: '/zh-cn/appendix/5-data/database-fundamentals'
+        },
+        {
+          text: '数据模型全景（文档 / 图 / 时序 / 向量）',
+          link: '/zh-cn/appendix/5-data/data-models'
+        },
+        {
+          text: '数据埋点与用户行为采集',
+          link: '/zh-cn/appendix/5-data/data-tracking'
+        },
+        {
+          text: '数据分析基础（统计 / 指标 / 漏斗）',
+          link: '/zh-cn/appendix/5-data/data-analysis'
+        },
+        {
+          text: 'A/B 测试与实验驱动',
+          link: '/zh-cn/appendix/5-data/ab-testing'
+        },
+        {
+          text: '数据可视化与仪表盘',
+          link: '/zh-cn/appendix/5-data/data-visualization'
+        },
+        {
+          text: '数据治理与数据质量',
+          link: '/zh-cn/appendix/5-data/data-governance'
+        }
+      ]
+    },
+    {
+      text: '六、架构与系统设计',
+      collapsed: false,
+      items: [
+        {
+          text: '从单体到微服务的演进',
+          link: '/zh-cn/appendix/6-architecture-and-system-design/monolith-to-microservices'
+        },
+        {
+          text: '分布式系统的挑战',
+          link: '/zh-cn/appendix/6-architecture-and-system-design/distributed-systems'
+        },
+        {
+          text: '高可用与容灾',
+          link: '/zh-cn/appendix/6-architecture-and-system-design/high-availability'
+        },
+        {
+          text: '系统设计方法论',
+          link: '/zh-cn/appendix/6-architecture-and-system-design/system-design-methodology'
+        }
+      ]
+    },
+    {
+      text: '七、基础设施与运维',
+      collapsed: false,
+      items: [
+        {
+          text: 'Linux 基础',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/linux-basics'
+        },
+        {
+          text: 'Docker 容器化',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/docker-containers'
+        },
+        {
+          text: 'Kubernetes 编排',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/kubernetes'
+        },
+        {
+          text: 'CI / CD 自动化',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/ci-cd'
+        },
+        {
+          text: '域名、DNS 与 HTTPS',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/dns-https'
+        },
+        {
+          text: '负载均衡与网关',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/load-balancing-gateway'
+        },
+        {
+          text: '网关与反向代理',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/gateway-proxy'
+        },
+        {
+          text: '云平台实战',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-platforms'
+        },
+        {
+          text: 'IAM 权限管理',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-iam'
+        },
+        {
+          text: '对象存储与 CDN',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/cloud-storage-cdn'
+        },
+        {
+          text: '基础设施即代码',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/infrastructure-as-code'
+        },
+        {
+          text: '监控、日志与告警',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/monitoring-logging'
+        },
+        {
+          text: '故障排查与应急响应',
+          link: '/zh-cn/appendix/7-infrastructure-and-operations/incident-response'
+        }
+      ]
+    },
+    {
+      text: '八、人工智能',
+      collapsed: false,
+      items: [
+        {
+          text: 'AI 简史与核心概念',
+          link: '/zh-cn/appendix/8-artificial-intelligence/ai-history'
+        },
+        {
+          text: '神经网络与深度学习',
+          link: '/zh-cn/appendix/8-artificial-intelligence/neural-networks'
+        },
+        {
+          text: 'Transformer 与注意力机制',
+          link: '/zh-cn/appendix/8-artificial-intelligence/transformer-attention'
+        },
+        {
+          text: '大语言模型的工作原理',
+          link: '/zh-cn/appendix/8-artificial-intelligence/llm-principles'
+        },
+        {
+          text: '提示词工程',
+          link: '/zh-cn/appendix/8-artificial-intelligence/prompt-engineering'
+        },
+        {
+          text: '上下文工程',
+          link: '/zh-cn/appendix/8-artificial-intelligence/context-engineering'
+        },
+        {
+          text: '多模态模型（视觉 / 音频 / 视频）',
+          link: '/zh-cn/appendix/8-artificial-intelligence/multimodal-models'
+        },
+        {
+          text: '图像生成原理',
+          link: '/zh-cn/appendix/8-artificial-intelligence/image-generation'
+        },
+        {
+          text: '语音合成与识别',
+          link: '/zh-cn/appendix/8-artificial-intelligence/speech-synthesis-recognition'
+        },
+        {
+          text: 'Embedding 与向量检索',
+          link: '/zh-cn/appendix/8-artificial-intelligence/embedding-vector-retrieval'
+        },
+        {
+          text: 'RAG 架构',
+          link: '/zh-cn/appendix/8-artificial-intelligence/rag'
+        },
+        {
+          text: 'AI Agent 与工具调用',
+          link: '/zh-cn/appendix/8-artificial-intelligence/ai-agents'
+        },
+        {
+          text: 'AI 协议（MCP & A2A）',
+          link: '/zh-cn/appendix/8-artificial-intelligence/ai-protocols'
+        },
+        {
+          text: '模型微调与部署',
+          link: '/zh-cn/appendix/8-artificial-intelligence/model-finetuning-deployment'
+        },
+        {
+          text: 'AI 原生应用设计',
+          link: '/zh-cn/appendix/8-artificial-intelligence/ai-native-app-design'
+        },
+        {
+          text: 'AI 能力词典',
+          link: '/zh-cn/appendix/8-artificial-intelligence/ai-capability-dictionary'
+        }
+      ]
+    },
+    {
+      text: '九、工程素养',
+      collapsed: false,
+      items: [
+        {
+          text: '代码质量与重构',
+          link: '/zh-cn/appendix/9-engineering-excellence/code-quality-refactoring'
+        },
+        {
+          text: '测试策略',
+          link: '/zh-cn/appendix/9-engineering-excellence/testing-strategies'
+        },
+        {
+          text: '设计模式',
+          link: '/zh-cn/appendix/9-engineering-excellence/design-patterns'
+        },
+        {
+          text: '安全思维与攻防基础',
+          link: '/zh-cn/appendix/9-engineering-excellence/security-thinking'
+        },
+        {
+          text: '技术文档写作',
+          link: '/zh-cn/appendix/9-engineering-excellence/technical-writing'
+        },
+        {
+          text: '开源协作',
+          link: '/zh-cn/appendix/9-engineering-excellence/open-source-collaboration'
+        },
+        {
+          text: '技术选型方法论',
+          link: '/zh-cn/appendix/9-engineering-excellence/technology-selection'
+        }
+      ]
+    }
+  ]
+}
